Refactor and improve Pixel Bot Maker application.
This commit addresses a comprehensive list of bug fixes and improvements identified in a code review.

Key changes include:
- Refactored the monolithic main.py into a structured `pixel_bot` package with `core`, `gui`, and `utils` sub-modules.
- Fixed a syntax error in `image_analyzer.py`.
- Added an execution depth check to prevent infinite loops in the bot's execution.
- Fixed a memory leak by ensuring loop counters are properly cleaned up on exceptions.
- Implemented a robust logging system using Python's `logging` module, which logs to both a file and the GUI.
- Added validation for step configurations to catch errors before execution.
- Implemented performance optimizations, including an image cache and an optimized screen capture routine.
- Added a unit test suite for the key-parsing logic and refactored the code to be testable in a headless environment.
