--- main.py
@@ -1,2771 +1,0 @@
-import tkinter as tk
-from tkinter import scrolledtext, colorchooser, filedialog, ttk
-import time
-import json
-import random
-import re
-
-# Import our custom modules
-import numpy as np
-import mss
-from pynput import keyboard
-import threading
-import pygetwindow as gw
-import os
-import cv2
-from PIL import Image, ImageTk
-from screen_capture import capture_screen
-from image_analyzer import find_color, find_image
-from automation import click_at, right_click_at, type_text, click_and_drag, scroll_wheel, press_key_combination
-from settings_manager import SettingsManager
-
-class App(tk.Tk):
-    def __init__(self):
-        super().__init__()
-        self.title("Pixel Bot")
-        self.settings_manager = SettingsManager()
-        self.geometry("800x660")
-
-        # --- Color Theme ---
-        self.sleek_blue_theme = {
-            "bg_color": "#0A192F",  # Deep Navy
-            "widget_bg_color": "#1E3A5F", # Darker Blue-Gray
-            "text_color": "#E6F1FF",  # Soft White/Light Blue
-            "accent_color": "#64FFDA", # Electric Blue/Mint Green for accents
-            "button_color": "#007BFF", # A more standard blue for buttons
-            "button_text_color": "#FFFFFF"
-        }
-        # For now, we'll just have one theme. We can add a light theme later if needed.
-        self.dark_theme = self.sleek_blue_theme
-        self.light_theme = self.sleek_blue_theme
-        self._configure_theme()
-        self.configure(bg=self.bg_color)
-
-        # --- App State ---
-        self.running = False
-        self.scan_job = None
-        self.hotkey_listener = None
-        self.action_sequence = []
-        self.variables = {} # For the new variable system
-        self.execution_stack = [] # (sequence, index)
-        self.conditional_loop_retry_counts = {}
-        self.step_retry_counts = {} # To track retries on a per-step basis
-        self.time_condition_executed = set() # To prevent re-triggering
-        self.loop_counters = {} # To track 'repeat x times' loops
-        self.hide_window_var = tk.BooleanVar(value=self.settings_manager.get_setting('hide_bot_default'))
-        self.dry_run_var = tk.BooleanVar(value=False)
-        self.target_window_title = tk.StringVar()
-        self.target_window_title.set("") # Set to empty string initially
-
-        # Ensure templates directory exists
-        if not os.path.exists("templates"):
-            os.makedirs("templates")
-
-        # --- Main Layout Frames ---
-        main_frame = ttk.Frame(self, padding="10")
-        main_frame.pack(fill="both", expand=True)
-        main_frame.columnconfigure(0, weight=1, uniform="group1")
-        main_frame.columnconfigure(1, weight=1, uniform="group1")
-        main_frame.rowconfigure(0, weight=1)
-
-        left_frame = ttk.Frame(main_frame)
-        left_frame.grid(row=0, column=0, sticky="nsew", padx=(0, 5))
-        left_frame.rowconfigure(0, weight=1)
-        left_frame.columnconfigure(0, weight=1)
-
-        right_frame = ttk.Frame(main_frame)
-        right_frame.grid(row=0, column=1, sticky="nsew", padx=(5, 0))
-        right_frame.rowconfigure(0, weight=1)
-        right_frame.columnconfigure(0, weight=1)
-
-        # --- WIDGET CREATION (Right Panel) ---
-        self.log_area = scrolledtext.ScrolledText(right_frame, width=45, height=10, relief=tk.FLAT, insertbackground=self.text_color)
-        self.log_area.grid(row=0, column=0, sticky="nsew")
-        right_frame.rowconfigure(0, weight=1)
-        right_frame.columnconfigure(0, weight=1)
-
-
-        # --- WIDGET CREATION (Left Panel) ---
-        self.style = ttk.Style()
-        self.style.theme_use('default')
-        self._apply_custom_styles()
-
-
-        notebook = ttk.Notebook(left_frame, style='TNotebook')
-        notebook.pack(expand=True, fill='both')
-
-        main_tab = ttk.Frame(notebook, padding="10")
-        settings_tab = ttk.Frame(notebook, padding="10")
-        notebook.add(main_tab, text='Main')
-        notebook.add(settings_tab, text='Settings')
-
-        # --- Main Tab Content ---
-        main_tab.columnconfigure(0, weight=1)
-        main_tab.rowconfigure(0, weight=1) # Make the sequence frame expandable
-
-        # --- Sequence Editor UI ---
-        sequence_frame = ttk.LabelFrame(main_tab, text="Action Sequence", padding="10")
-        sequence_frame.grid(row=0, column=0, sticky="nsew", pady=(0, 10))
-        sequence_frame.columnconfigure(0, weight=1)
-        sequence_frame.rowconfigure(1, weight=1)
-
-        # Frame for Save/Load buttons
-        file_io_frame = ttk.Frame(sequence_frame)
-        file_io_frame.grid(row=0, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        ttk.Button(file_io_frame, text="Load Sequence", command=self.load_sequence).pack(side="left", padx=(0,5))
-        ttk.Button(file_io_frame, text="Save Sequence", command=self.save_sequence).pack(side="left")
-
-        list_container = ttk.Frame(sequence_frame)
-        list_container.grid(row=1, column=0, columnspan=2, sticky="nsew")
-        list_container.columnconfigure(0, weight=1)
-        list_container.rowconfigure(0, weight=1)
-
-        self.sequence_listbox = tk.Listbox(list_container, bg=self.widget_bg_color, fg=self.text_color, relief=tk.FLAT, height=10)
-        self.sequence_listbox.grid(row=0, column=0, sticky="nsew")
-        self.sequence_listbox.bind("<<ListboxSelect>>", self.on_sequence_select)
-
-        seq_button_frame = ttk.Frame(list_container)
-        seq_button_frame.grid(row=0, column=1, sticky="ns", padx=(5,0))
-        self.add_step_button = ttk.Button(seq_button_frame, text="Add", command=self.add_step, state=tk.DISABLED)
-        self.add_step_button.pack(pady=2, fill="x")
-        self.edit_step_button = ttk.Button(seq_button_frame, text="Edit", command=self.edit_step, state=tk.DISABLED)
-        self.edit_step_button.pack(pady=2, fill="x")
-        self.move_up_button = ttk.Button(seq_button_frame, text="Move Up", command=self.move_step_up, state=tk.DISABLED)
-        self.move_up_button.pack(pady=2, fill="x")
-        self.move_down_button = ttk.Button(seq_button_frame, text="Move Down", command=self.move_step_down, state=tk.DISABLED)
-        self.move_down_button.pack(pady=2, fill="x")
-        self.remove_step_button = ttk.Button(seq_button_frame, text="Remove", command=self.remove_step, state=tk.DISABLED)
-        self.remove_step_button.pack(pady=2, fill="x")
-
-        # --- Final Controls ---
-        controls_frame = ttk.LabelFrame(main_tab, text="Global Target", padding="10")
-        controls_frame.grid(row=1, column=0, sticky="ew", pady=(0, 10))
-        self.target_window_label = ttk.Label(controls_frame, textvariable=self.target_window_title, wraplength=380, justify="left", style="Card.TLabel")
-        self.target_window_label.pack(pady=5, fill="x", expand=True, ipady=5)
-        ttk.Button(controls_frame, text="Change Target Window", command=self.prompt_for_window_selection).pack(pady=(10,5))
-
-        # --- Most Loaded Sequences ---
-        most_loaded_frame = ttk.LabelFrame(main_tab, text="Frequently Used", padding="10")
-        most_loaded_frame.grid(row=2, column=0, sticky="ew")
-        self.most_loaded_container = ttk.Frame(most_loaded_frame)
-        self.most_loaded_container.pack(fill="x", pady=5)
-
-        bot_controls_frame = ttk.Frame(main_tab)
-        bot_controls_frame.grid(row=3, column=0, sticky="ew", pady=(10, 0))
-        bot_controls_frame.columnconfigure(0, weight=1)
-
-        self.hide_window_check = ttk.Checkbutton(bot_controls_frame, text="Hide window when bot is running", variable=self.hide_window_var)
-        self.hide_window_check.pack()
-
-        self.dry_run_check = ttk.Checkbutton(bot_controls_frame, text="Dry Run (log actions without executing)", variable=self.dry_run_var)
-        self.dry_run_check.pack(pady=5)
-
-        self.start_button = ttk.Button(bot_controls_frame, text="Start Bot", command=self.toggle_bot, style="Accent.TButton")
-        self.start_button.pack(pady=10)
-
-        # --- Settings Tab Content ---
-        settings_content_frame = ttk.Frame(settings_tab)
-        settings_content_frame.pack(fill="both", expand=True)
-
-        # --- General Settings ---
-        general_frame = ttk.LabelFrame(settings_content_frame, text="General", padding="10")
-        general_frame.pack(fill="x", pady=5, anchor="n")
-
-        self.hide_bot_default_var = tk.BooleanVar(value=self.settings_manager.get_setting('hide_bot_default'))
-        hide_bot_check = ttk.Checkbutton(general_frame, text="Hide window by default when bot is running", variable=self.hide_bot_default_var, command=self.save_hide_bot_default)
-        hide_bot_check.pack(anchor="w", padx=5, pady=2)
-
-        # --- Appearance Settings ---
-        # Theme selection is removed as we now have a single, unified theme.
-
-        # --- Image Matching Settings ---
-        matching_frame = ttk.LabelFrame(settings_content_frame, text="Image Matching", padding="10")
-        matching_frame.pack(fill="x", pady=5, anchor="n")
-
-        self.similarity_threshold_var = tk.DoubleVar(value=self.settings_manager.get_setting('image_similarity_threshold'))
-
-        threshold_inner_frame = ttk.Frame(matching_frame)
-        threshold_inner_frame.pack(fill="x", pady=2)
-
-        ttk.Label(threshold_inner_frame, text="Similarity Threshold:").pack(side="left", padx=5)
-        self.similarity_label_var = tk.StringVar(value=f"{self.similarity_threshold_var.get():.2f}")
-        ttk.Label(threshold_inner_frame, textvariable=self.similarity_label_var, style="Card.TLabel", width=5).pack(side="left")
-
-        similarity_slider = ttk.Scale(
-            threshold_inner_frame,
-            from_=0.0,
-            to=1.0,
-            orient=tk.HORIZONTAL,
-            variable=self.similarity_threshold_var,
-            command=self.save_similarity_threshold,
-        )
-        similarity_slider.pack(side="left", fill="x", expand=True, padx=5)
-
-
-        # --- Hotkey Settings ---
-        hotkey_frame = ttk.LabelFrame(settings_content_frame, text="Hotkey", padding="10")
-        hotkey_frame.pack(fill="x", pady=5, anchor="n")
-
-        hotkey_inner_frame = ttk.Frame(hotkey_frame)
-        hotkey_inner_frame.pack(fill="x", pady=2)
-        ttk.Label(hotkey_inner_frame, text="Start/Stop Hotkey:").pack(side="left", padx=5)
-        self.hotkey_label_var = tk.StringVar(value=self.settings_manager.get_setting('hotkey'))
-        ttk.Label(hotkey_inner_frame, textvariable=self.hotkey_label_var, style="Card.TLabel", padding=(10, 5), width=12, anchor="center").pack(side="left")
-        self.change_hotkey_button = ttk.Button(hotkey_inner_frame, text="Change...", command=self.change_hotkey)
-        self.change_hotkey_button.pack(side="left", padx=5)
-
-        # --- Default Wait Times UI ---
-        self.wait_frame = ttk.LabelFrame(settings_content_frame, text="Default Post-Action Wait", padding="10")
-        self.wait_frame.pack(fill="x", pady=5, anchor="n")
-
-        default_wait_settings = self.settings_manager.get_setting('default_wait_times')
-        self.default_wait_type = tk.StringVar(value=default_wait_settings.get('type', 'Fixed'))
-        self.default_fixed_wait = tk.StringVar(value=str(default_wait_settings.get('fixed_time', 1)))
-        self.default_min_wait = tk.StringVar(value=str(default_wait_settings.get('min_time', 1)))
-        self.default_max_wait = tk.StringVar(value=str(default_wait_settings.get('max_time', 2)))
-
-        wait_type_frame = ttk.Frame(self.wait_frame)
-        ttk.Radiobutton(wait_type_frame, text="Fixed", variable=self.default_wait_type, value="Fixed", command=self.on_default_wait_type_change).pack(side="left")
-        ttk.Radiobutton(wait_type_frame, text="Random", variable=self.default_wait_type, value="Random", command=self.on_default_wait_type_change).pack(side="left")
-        wait_type_frame.pack(fill="x", pady=(0,5))
-
-        self.default_fixed_wait_frame = ttk.Frame(self.wait_frame)
-        ttk.Label(self.default_fixed_wait_frame, text="Default Fixed Wait (sec):").pack(side="left", padx=5)
-        ttk.Entry(self.default_fixed_wait_frame, textvariable=self.default_fixed_wait, width=7).pack(side="left")
-
-        self.default_random_wait_frame = ttk.Frame(self.wait_frame)
-        ttk.Label(self.default_random_wait_frame, text="Min (sec):").pack(side="left", padx=5)
-        ttk.Entry(self.default_random_wait_frame, textvariable=self.default_min_wait, width=7).pack(side="left")
-        ttk.Label(self.default_random_wait_frame, text="Max (sec):").pack(side="left", padx=(10,5))
-        ttk.Entry(self.default_random_wait_frame, textvariable=self.default_max_wait, width=7).pack(side="left")
-
-        ttk.Button(self.wait_frame, text="Save Default Waits", command=self.save_default_wait_times).pack(pady=(10,0))
-
-        self.on_default_wait_type_change()
-
-        self.log("Welcome! Please select a target window to begin.")
-        try:
-            # Attempt to get local timezone for user-friendliness
-            local_tz = time.tzname[time.daylight] if time.daylight and len(time.tzname) > 1 else time.tzname[0]
-            self.log(f"Bot's current time is {time.strftime('%Y-%m-%d %H:%M:%S')} ({local_tz}). Time conditions use the bot's system clock.")
-        except Exception:
-            # Fallback for environments where tzname might not be available
-            self.log(f"Bot's current time is {time.strftime('%Y-%m-%d %H:%M:%S')}. Time conditions use the bot's system clock.")
-        self.update_most_loaded_list()
-        self.after(200, lambda: self.prompt_for_window_selection(is_splash=True))
-
-    def _configure_theme(self):
-        theme = self.settings_manager.get_setting('theme')
-        colors = self.dark_theme if theme == 'dark' else self.light_theme
-
-        self.bg_color = colors['bg_color']
-        self.widget_bg_color = colors['widget_bg_color']
-        self.text_color = colors['text_color']
-        self.button_color = colors['button_color']
-        self.button_text_color = colors['button_text_color']
-
-    def _apply_custom_styles(self):
-        theme = self.settings_manager.get_setting('theme')
-        colors = self.dark_theme if theme == 'dark' else self.light_theme
-
-        # General widget styling
-        self.style.configure('TFrame', background=colors['bg_color'])
-        self.style.configure('TLabel', background=colors['bg_color'], foreground=colors['text_color'], padding=5)
-        self.style.configure('TButton', background=colors['widget_bg_color'], foreground=colors['text_color'], padding=5)
-        self.style.map('TButton', background=[('active', colors['button_color'])])
-        self.style.configure('TCheckbutton', background=colors['bg_color'], foreground=colors['text_color'])
-        self.style.map('TCheckbutton', background=[('active', colors['bg_color'])])
-        self.style.configure('TRadiobutton', background=colors['bg_color'], foreground=colors['text_color'])
-        self.style.map('TRadiobutton', background=[('active', colors['bg_color'])])
-        self.style.configure('TEntry', fieldbackground=colors['widget_bg_color'], foreground=colors['text_color'], insertcolor=colors['text_color'])
-        self.style.configure('TLabelFrame', background=colors['bg_color'], relief=tk.SOLID, borderwidth=1)
-        self.style.configure('TLabelFrame.Label', background=colors['bg_color'], foreground=colors['text_color'])
-
-
-        # Special "Card" style for labels that need a background
-        self.style.configure("Card.TLabel", background=colors['widget_bg_color'], relief=tk.SOLID, borderwidth=1)
-
-        # Special "Accent" button style
-        self.style.configure("Accent.TButton", background=colors['accent_color'], foreground=colors['bg_color'])
-        self.style.map("Accent.TButton", background=[('active', colors['text_color'])])
-
-        # Notebook styling
-        self.style.configure('TNotebook', background=colors['bg_color'], borderwidth=0)
-        self.style.configure('TNotebook.Tab', background=colors['bg_color'], foreground=colors['text_color'], padding=[10, 5], font=("Segoe UI", 10))
-        self.style.map('TNotebook.Tab', background=[('selected', colors['widget_bg_color'])], foreground=[('selected', colors['accent_color'])])
-
-        # Apply background and foreground to the log area specifically
-        self.log_area.config(bg=colors['widget_bg_color'], fg=colors['text_color'])
-
-
-    def prompt_for_window_selection(self, is_splash=False):
-        self.log("Prompting for target window selection...")
-        WindowSelector(self, is_splash=is_splash)
-
-    def on_window_selected(self, title):
-        self.target_window_title.set(title)
-        self.add_step_button.state(['!disabled'])
-        self.log(f"Global target window set to: {title}")
-
-    def on_sequence_select(self, event):
-        selected_indices = self.sequence_listbox.curselection()
-        if selected_indices:
-            index = selected_indices[0]
-            self.edit_step_button.state(['!disabled'])
-            self.remove_step_button.state(['!disabled'])
-
-            # Enable/disable Move Up button
-            if index > 0:
-                self.move_up_button.state(['!disabled'])
-            else:
-                self.move_up_button.state(['disabled'])
-
-            # Enable/disable Move Down button
-            if index < len(self.action_sequence) - 1:
-                self.move_down_button.state(['!disabled'])
-            else:
-                self.move_down_button.state(['disabled'])
-        else:
-            self.edit_step_button.state(['disabled'])
-            self.remove_step_button.state(['disabled'])
-            self.move_up_button.state(['disabled'])
-            self.move_down_button.state(['disabled'])
-
-    def add_step(self):
-        StepEditor(self)
-
-    def edit_step(self):
-        selected_indices = self.sequence_listbox.curselection()
-        if not selected_indices:
-            return
-        index = selected_indices[0]
-        step_data = self.action_sequence[index]
-        StepEditor(self, step_data=step_data, index=index)
-
-    def remove_step(self):
-        selected_indices = self.sequence_listbox.curselection()
-        if not selected_indices:
-            return
-        index = selected_indices[0]
-        self.action_sequence.pop(index)
-        self.update_sequence_listbox()
-        self.log(f"Removed step {index+1}.")
-
-    def move_step_up(self):
-        selected_indices = self.sequence_listbox.curselection()
-        if not selected_indices:
-            return
-
-        index = selected_indices[0]
-        if index > 0:
-            # Swap with the previous item
-            self.action_sequence[index], self.action_sequence[index - 1] = self.action_sequence[index - 1], self.action_sequence[index]
-            self.update_sequence_listbox()
-            # Reselect the item at its new position
-            self.sequence_listbox.selection_set(index - 1)
-            self.on_sequence_select(None) # Update button states
-
-    def move_step_down(self):
-        selected_indices = self.sequence_listbox.curselection()
-        if not selected_indices:
-            return
-
-        index = selected_indices[0]
-        if index < len(self.action_sequence) - 1:
-            # Swap with the next item
-            self.action_sequence[index], self.action_sequence[index + 1] = self.action_sequence[index + 1], self.action_sequence[index]
-            self.update_sequence_listbox()
-            # Reselect the item at its new position
-            self.sequence_listbox.selection_set(index + 1)
-            self.on_sequence_select(None) # Update button states
-
-    def on_step_saved(self, step_data, index=None):
-        if index is not None:
-            self.action_sequence[index] = step_data
-        else:
-            self.action_sequence.append(step_data)
-        self.update_sequence_listbox()
-
-    def update_sequence_listbox(self):
-        self.sequence_listbox.delete(0, tk.END)
-        for i, step in enumerate(self.action_sequence):
-            step_type = step.get('step_type', 'simple')
-            action_type = step.get('action_type') # Get action_type for simple steps
-            text = f"{i+1}: "
-
-            if step_type == 'simple':
-                if action_type == 'Set Variable':
-                    params = step.get('action_params', {})
-                    text += f"Set Var: '{params.get('variable_name', 'N/A')}' = '{params.get('variable_value', 'N/A')}'"
-                elif action_type == 'Modify Variable':
-                    params = step.get('action_params', {})
-                    text += f"Modify Var: '{params.get('modify_variable_name', 'N/A')}' {params.get('modify_variable_operation', '?')} {params.get('modify_variable_value', '?')}"
-                elif action_type == 'OCR':
-                    params = step.get('action_params', {})
-                    text += f"OCR to Var: '{params.get('output_variable_name', 'N/A')}'"
-                else:
-                    mode = step.get('detection_mode', '?')
-                    action = step.get('action_type', '?')
-                    target = step.get('detection_target_name', 'Unknown')
-                    text += f"Find {mode} '{target}', then {action}"
-            elif step_type == 'conditional_loop':
-                primary_target_name = step.get('primary_target', {}).get('detection_target_name', 'N/A')
-                fallback_action = step.get('on_fail', {}).get('action_type', 'N/A')
-                retries = step.get('max_retries', 'N/A')
-                text += f"CONDITIONAL ({retries}x): Find '{primary_target_name}', on fail: {fallback_action}"
-            elif step_type == 'loop':
-                loop_mode = step.get('loop_mode', 'repeat')
-                if loop_mode == 'repeat':
-                    repeat_count = step.get('loop_repeat_count', 'N/A')
-                    text += f"LOOP: Repeat {repeat_count} times"
-                else:
-                    text += f"LOOP: Until condition"
-            elif step_type == 'time_based_condition':
-                time_cond = step.get('time_condition', {})
-                hour = time_cond.get('hour', '??')
-                minute = time_cond.get('minute', '??')
-                text += f"TIME CONDITION: at {hour:02d}:{minute:02d}"
-            else:
-                text += "Unknown Step Type"
-
-            self.sequence_listbox.insert(tk.END, text)
-        self.on_sequence_select(None)
-
-    def save_sequence(self):
-        if not self.action_sequence:
-            self.log("Cannot save an empty sequence.")
-            return
-
-        filepath = filedialog.asksaveasfilename(
-            parent=self,
-            title="Save Sequence",
-            filetypes=(("JSON files", "*.json"), ("All files", "*.*")),
-            defaultextension=".json"
-        )
-        if not filepath:
-            return
-
-        try:
-            with open(filepath, 'w') as f:
-                json.dump(self.action_sequence, f, indent=4)
-            self.log(f"Sequence saved to {os.path.basename(filepath)}")
-        except Exception as e:
-            self.log(f"Error saving sequence: {e}")
-
-    def update_most_loaded_list(self):
-        # Clear existing labels
-        for widget in self.most_loaded_container.winfo_children():
-            widget.destroy()
-        most_loaded = self.settings_manager.get_most_loaded_sequences(count=3)
-
-        if not most_loaded:
-            tk.Label(self.most_loaded_container, text="No sequences loaded yet.", bg=self.bg_color, fg=self.text_color, anchor="w").pack(fill="x", pady=2)
-            return
-
-        for filepath, count in most_loaded:
-            # Shorten the display name
-            display_name = os.path.basename(filepath)
-
-            # Create a clickable label
-            label_text = f"  {display_name} ({count} loads)"
-            label = ttk.Label(self.most_loaded_container, text=label_text, style="Card.TLabel", anchor="w", cursor="hand2")
-            label.pack(fill="x", pady=(0, 2))
-            label.bind("<Button-1>", lambda e, path=filepath: self.load_sequence(path))
-
-
-    def load_sequence(self, filepath=None):
-        if filepath is None:
-            filepath = filedialog.askopenfilename(
-                parent=self,
-                title="Load Sequence",
-                filetypes=(("JSON files", "*.json"), ("All files", "*.*"))
-            )
-        if not filepath:
-            return
-
-        try:
-            with open(filepath, 'r') as f:
-                loaded_sequence = json.load(f)
-
-            if not isinstance(loaded_sequence, list):
-                raise TypeError("File does not contain a valid sequence list.")
-
-            self.action_sequence = loaded_sequence
-            self.update_sequence_listbox()
-            self.log(f"Sequence loaded from {os.path.basename(filepath)}")
-
-            # Increment load count and update the list
-            self.settings_manager.increment_sequence_load_count(filepath)
-            self.update_most_loaded_list()
-        except json.JSONDecodeError:
-            self.log("Error: The selected file is not a valid JSON file.")
-        except TypeError as e:
-            self.log(f"Error loading sequence: {e}")
-        except Exception as e:
-            self.log(f"An unexpected error occurred while loading: {e}")
-
-    def save_hide_bot_default(self):
-        self.settings_manager.set_setting('hide_bot_default', self.hide_bot_default_var.get())
-        self.log("Default 'Hide Bot' setting saved.")
-
-    def save_similarity_threshold(self, value):
-        """Called when the similarity slider is moved."""
-        # The 'value' argument is a string from the Scale widget, so we convert it to float
-        new_threshold = float(value)
-        self.settings_manager.set_setting('image_similarity_threshold', new_threshold)
-        self.similarity_label_var.set(f"{new_threshold:.2f}")
-        # No log message here to prevent spamming the log while dragging the slider
-
-    def apply_theme(self):
-        theme = self.theme_var.get()
-        self.settings_manager.set_setting('theme', theme)
-        self.log("Theme changed. Please restart the application for the changes to take full effect.")
-
-    def on_default_wait_type_change(self):
-        wait_type = self.default_wait_type.get()
-        if wait_type == "Fixed":
-            self.default_fixed_wait_frame.pack(fill="x", pady=2)
-            self.default_random_wait_frame.pack_forget()
-        elif wait_type == "Random":
-            self.default_fixed_wait_frame.pack_forget()
-            self.default_random_wait_frame.pack(fill="x", pady=2)
-        else: # Should not happen with radio buttons
-            self.default_fixed_wait_frame.pack_forget()
-            self.default_random_wait_frame.pack_forget()
-
-    def save_default_wait_times(self):
-        try:
-            new_settings = {
-                "type": self.default_wait_type.get(),
-                "fixed_time": float(self.default_fixed_wait.get()),
-                "min_time": float(self.default_min_wait.get()),
-                "max_time": float(self.default_max_wait.get())
-            }
-            if new_settings['min_time'] >= new_settings['max_time']:
-                self.log("Error: Min wait time must be less than max wait time.")
-                return
-
-            self.settings_manager.set_setting('default_wait_times', new_settings)
-            self.log("Default wait times saved successfully.")
-        except ValueError:
-            self.log("Error: Wait times must be valid numbers.")
-
-    def change_hotkey(self):
-        HotkeyChangeDialog(self)
-
-    def start_hotkey_listener(self):
-        if self.hotkey_listener:
-            self.hotkey_listener.stop()
-
-        hotkey_str = self.settings_manager.get_setting('hotkey')
-        target_key = None
-
-        try:
-            if hasattr(keyboard.Key, hotkey_str.lower()):
-                target_key = getattr(keyboard.Key, hotkey_str.lower())
-            elif len(hotkey_str) == 1:
-                target_key = keyboard.KeyCode.from_char(hotkey_str)
-        except Exception as e:
-            self.log(f"Error parsing hotkey '{hotkey_str}': {e}. Defaulting to F9.")
-            target_key = keyboard.Key.f9
-            self.hotkey_label_var.set("f9")
-            self.settings_manager.set_setting('hotkey', "f9")
-
-        if target_key is None:
-            self.log(f"Invalid hotkey '{hotkey_str}'. Please set a valid key. Defaulting to F9.")
-            target_key = keyboard.Key.f9
-            self.hotkey_label_var.set("f9")
-            self.settings_manager.set_setting('hotkey', "f9")
-
-        def on_press(key):
-            if key == target_key:
-                self.after(0, self.toggle_bot)
-
-        self.hotkey_listener = keyboard.Listener(on_press=on_press)
-        self.hotkey_listener.start()
-        self.log(f"Bot running... Press '{hotkey_str.upper()}' to stop.")
-
-    def toggle_bot(self):
-        if self.running:
-            self.running = False
-            self.start_button.config(text="Start Bot")
-            if self.scan_job:
-                self.after_cancel(self.scan_job)
-            self.scan_job = None
-            if self.hotkey_listener:
-                self.hotkey_listener.stop()
-            self.hotkey_listener = None
-            self.log("Bot stopped by user.")
-            if self.hide_window_var.get():
-                self.deiconify()
-        else:
-            if not self.action_sequence:
-                self.log("Cannot start: Action sequence is empty.")
-                return
-            self.variables.clear()
-            self.execution_stack = [(self.action_sequence, 0)]
-            self.conditional_loop_retry_counts.clear() # Reset all conditional loop counters
-            self.step_retry_counts.clear() # Reset step-specific retries
-            self.time_condition_executed.clear() # Reset executed time conditions
-            self.loop_counters.clear() # Reset all loop counters
-            self.running = True
-            self.start_button.config(text="Stop Bot")
-            self.start_hotkey_listener()
-            if self.hide_window_var.get():
-                self.withdraw()
-            self.run_scan_loop()
-
-    def _handle_post_action_wait(self, step):
-        wait_params = step.get('wait_params', {})
-        wait_type = wait_params.get('type', 'Fixed')
-        wait_duration = 0
-
-        if wait_type == 'Fixed':
-            wait_duration = wait_params.get('fixed_time', 1.0)
-        elif wait_type == 'Random':
-            min_time = wait_params.get('min_time', 1.0)
-            max_time = wait_params.get('max_time', 2.0)
-            wait_duration = random.uniform(min_time, max_time)
-
-        # The wait duration from config is in seconds. 'after' needs milliseconds.
-        wait_ms = int(wait_duration * 1000)
-
-        if wait_ms > 0:
-            self.log(f"Waiting for {wait_duration:.2f} seconds...")
-            self.scan_job = self.after(wait_ms, self.run_scan_loop)
-        else:
-            # If no wait, or wait is 0, proceed to the next step immediately,
-            # but use after(10) to yield to the main GUI loop and prevent freezing.
-            self.scan_job = self.after(10, self.run_scan_loop)
-
-    def _substitute_variables(self, text):
-        """
-        Substitutes placeholders like {{var_name}} in a string with their
-        values from the self.variables dictionary.
-        """
-        if not isinstance(text, str):
-            return text
-
-        # This regex finds all occurrences of {{...}}
-        return re.sub(r'\{\{(.*?)\}\}', self._replace_match, text)
-
-    def _replace_match(self, match):
-        """
-        Helper function for re.sub to look up the variable.
-        """
-        var_name = match.group(1).strip()
-        # Return the value if found, otherwise return the original placeholder
-        return self.variables.get(var_name, match.group(0))
-
-    def _evaluate_condition(self, condition):
-        var_name = condition.get('variable', '').strip()
-        operator = condition.get('operator')
-        value_to_compare = self._substitute_variables(condition.get('value', ''))
-
-        if not var_name or not operator:
-            self.log(f"Invalid condition: {condition}")
-            return False
-
-        actual_value = self._substitute_variables(var_name)
-
-        # For numeric comparisons, try to convert both to floats
-        try:
-            num_actual_value = float(actual_value)
-            num_value_to_compare = float(value_to_compare)
-            is_numeric = True
-        except (ValueError, TypeError):
-            is_numeric = False
-
-        if operator == "equals":
-            return actual_value == value_to_compare
-        elif operator == "not equals":
-            return actual_value != value_to_compare
-        elif operator == "contains":
-            return value_to_compare in actual_value
-        elif operator == "not contains":
-            return value_to_compare not in actual_value
-        elif is_numeric:
-            if operator == "is greater than":
-                return num_actual_value > num_value_to_compare
-            elif operator == "is less than":
-                return num_actual_value < num_value_to_compare
-
-        self.log(f"Unsupported operator '{operator}' for non-numeric comparison.")
-        return False
-
-
-    def run_scan_loop(self):
-        if not self.running:
-            return
-
-        if not self.execution_stack:
-            self.log("Action sequence complete. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        current_sequence, current_index = self.execution_stack[-1]
-
-        if current_index >= len(current_sequence):
-            self.execution_stack.pop()
-            self.scan_job = self.after(10, self.run_scan_loop) # Continue processing parent sequence
-            return
-
-        current_step = current_sequence[current_index]
-        step_type = current_step.get('step_type', 'simple')
-        action_type = current_step.get('action_type')
-        step_number_str = f"Step {current_index + 1}" # For logging
-
-        # Handle non-UI actions first
-        if action_type == 'Set Variable':
-            params = current_step.get('action_params', {})
-            var_name = params.get('variable_name')
-            var_value = self._substitute_variables(params.get('variable_value')) # Allow variables in values
-
-            if not var_name:
-                self.log(f"Error in {step_number_str}: 'Set Variable' action has no variable name. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            if self.dry_run_var.get():
-                self.log(f"[DRY RUN] Would set variable '{var_name}' to '{var_value}'")
-            else:
-                self.log(f"Setting variable '{var_name}' to '{var_value}'")
-                self.variables[var_name] = var_value
-
-            self.execution_stack[-1] = (current_sequence, current_index + 1)
-            self._handle_post_action_wait(current_step) # Still respect wait times
-            return
-        elif action_type == 'Modify Variable':
-            params = current_step.get('action_params', {})
-            var_name = params.get('modify_variable_name')
-            operation = params.get('modify_variable_operation')
-            value_str = self._substitute_variables(params.get('modify_variable_value', '0'))
-
-            if not var_name:
-                self.log(f"Error in {step_number_str}: 'Modify Variable' action has no variable name. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            current_value_str = self.variables.get(var_name, '0')
-
-            if self.dry_run_var.get():
-                self.log(f"[DRY RUN] Would {operation} variable '{var_name}' by '{value_str}'")
-            else:
-                if operation == 'set':
-                    self.variables[var_name] = value_str
-                    self.log(f"Set variable '{var_name}' to '{value_str}'")
-                else: # add or subtract
-                    try:
-                        current_value = float(current_value_str)
-                        value_to_op = float(value_str)
-                        new_value = 0
-                        if operation == 'add':
-                            new_value = current_value + value_to_op
-                        elif operation == 'subtract':
-                            new_value = current_value - value_to_op
-
-                        # Store as int if it's a whole number, otherwise float
-                        if new_value == int(new_value):
-                            self.variables[var_name] = str(int(new_value))
-                        else:
-                            self.variables[var_name] = str(new_value)
-
-                        self.log(f"Variable '{var_name}' {operation}ed by {value_to_op}. New value: {self.variables[var_name]}")
-
-                    except ValueError:
-                        self.log(f"Error in {step_number_str}: Cannot perform arithmetic on non-numeric variable '{var_name}' (value: '{current_value_str}') or input '{value_str}'. Stopping bot.")
-                        self.toggle_bot()
-                        return
-
-            self.execution_stack[-1] = (current_sequence, current_index + 1)
-            self._handle_post_action_wait(current_step)
-            return
-        elif action_type == 'OCR':
-            params = current_step.get('action_params', {})
-            region = params.get('ocr_region')
-            output_var = params.get('output_variable_name')
-
-            if not region or not output_var:
-                self.log(f"Error in {step_number_str}: OCR step is not configured correctly. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            if self.dry_run_var.get():
-                self.log(f"[DRY RUN] Would perform OCR on region {region} and save to '{output_var}'")
-                # In a dry run, we can't know the result, so we'll save a placeholder
-                self.variables[output_var] = "DRY_RUN_OCR_RESULT"
-            else:
-                self.log(f"Performing OCR on region {region} and saving to '{output_var}'...")
-                screenshot = capture_screen(region)
-                from image_analyzer import extract_text_from_image
-                extracted_text = extract_text_from_image(screenshot)
-
-                if extracted_text == "TESSERACT_NOT_FOUND":
-                    self.log("FATAL: Tesseract OCR engine not found. Please install it to use the OCR feature. Stopping bot.")
-                    self.toggle_bot()
-                    return
-
-                self.log(f"OCR Result: '{extracted_text}'. Stored in variable '{output_var}'.")
-                self.variables[output_var] = extracted_text
-
-            self.execution_stack[-1] = (current_sequence, current_index + 1)
-            self._handle_post_action_wait(current_step)
-            return
-
-        # For other actions, we pass the step and its context
-        step_context = {'sequence': current_sequence, 'index': current_index, 'number_str': step_number_str}
-
-        if step_type == 'simple':
-            self._execute_simple_step(current_step, step_context)
-        elif step_type == 'conditional_branch':
-            self.log(f"Evaluating condition for {step_number_str}...")
-            # First, advance the parent sequence *before* pushing a new one.
-            self.execution_stack[-1] = (current_sequence, current_index + 1)
-
-            if self._evaluate_condition(current_step.get('condition')):
-                self.log("Condition is TRUE. Executing IF branch.")
-                if_branch = current_step.get('if_branch', [])
-                if if_branch:
-                    self.execution_stack.append((if_branch, 0))
-            else:
-                self.log("Condition is FALSE. Executing ELSE branch.")
-                else_branch = current_step.get('else_branch', [])
-                if else_branch:
-                    self.execution_stack.append((else_branch, 0))
-
-            # Continue to the next loop iteration immediately to process the new stack
-            self.scan_job = self.after(10, self.run_scan_loop)
-            return
-        elif step_type == 'loop':
-            self._execute_loop_step(current_step, step_context)
-        elif step_type == 'conditional_loop':
-            self._execute_conditional_loop_step(current_step, step_context)
-        elif step_type == 'time_based_condition':
-            self._execute_time_based_condition_step(current_step, step_context)
-        else:
-            self.log(f"Error: Unknown step type '{step_type}' at {step_number_str}. Stopping bot.")
-            self.toggle_bot()
-
-    def _execute_loop_step(self, step, context):
-        """
-        Handles the execution of a 'loop' step by managing the execution stack
-        instead of using a blocking internal loop.
-        """
-        loop_mode = step.get('loop_mode', 'repeat')
-        step_id = id(step)
-
-        # --- REPEAT X TIMES mode ---
-        if loop_mode == 'repeat':
-            repeat_count = step.get('loop_repeat_count', 1)
-            # Get current iteration count, default to 0 if not present
-            current_iteration = self.loop_counters.get(step_id, 0)
-
-            if current_iteration < repeat_count:
-                self.log(f"{context['number_str']} (Loop): Running iteration {current_iteration + 1}/{repeat_count}.")
-                # Increment the counter for the next time we see this step
-                self.loop_counters[step_id] = current_iteration + 1
-
-                # 1. Push the current loop step back onto the stack. When the sub-actions
-                #    are done, the bot will re-evaluate this same step.
-                self.execution_stack.append((context['sequence'], context['index']))
-                # 2. Push the sub-actions onto the stack to be executed now.
-                self.execution_stack.append((step.get('loop_actions', []), 0))
-            else:
-                # Loop is finished
-                self.log(f"{context['number_str']} (Loop): Finished {repeat_count} iterations.")
-                # Clean up the counter
-                if step_id in self.loop_counters:
-                    del self.loop_counters[step_id]
-                # Advance the parent sequence to the *next* step
-                self.execution_stack[-1] = (context['sequence'], context['index'] + 1)
-
-            # Immediately schedule the next scan to process the updated stack
-            self.scan_job = self.after(10, self.run_scan_loop)
-            return
-
-        # --- UNTIL CONDITION MET mode ---
-        elif loop_mode == 'until':
-            max_retries = step.get('max_retries', 10)
-            current_retries = self.loop_counters.get(step_id, 0)
-
-            if current_retries >= max_retries:
-                self.log(f"Loop failed: Condition not met after {max_retries} retries. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            # Get window and region details first
-            try:
-                target_window_title = step.get("window_title")
-                target_windows = gw.getWindowsWithTitle(target_window_title)
-                if not target_windows:
-                    self.log(f"{context['number_str']} (Loop): Window '{target_window_title}' not found. Re-scanning...")
-                    self.scan_job = self.after(2000, self.run_scan_loop)
-                    return
-                target_window = target_windows[0]
-                scan_region = {'top': target_window.top, 'left': target_window.left, 'width': target_window.width, 'height': target_window.height}
-                if step.get('search_region'):
-                    custom_region = step['search_region']
-                    scan_region = {'top': target_window.top + custom_region['y'], 'left': target_window.left + custom_region['x'], 'width': custom_region['width'], 'height': custom_region['height']}
-            except Exception as e:
-                self.log(f"Error getting window details: {e}. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            # 1. Check for the condition
-            self.log(f"Executing {context['number_str']} (Loop): Checking for condition '{step.get('loop_condition_target_name')}'. Attempt {current_retries + 1}/{max_retries}.")
-            haystack_img = capture_screen(scan_region)
-            condition_found = False
-            try:
-                targets = step['loop_condition_target']
-                if isinstance(targets, str): targets = [targets]
-                needle_imgs = [cv2.imread(p, cv2.IMREAD_UNCHANGED) for p in targets]
-                threshold = self.settings_manager.get_setting('image_similarity_threshold')
-                if find_image(haystack_img, [img for img in needle_imgs if img is not None], threshold=threshold):
-                    condition_found = True
-            except Exception as e:
-                self.log(f"  - Error loading condition image(s): {e}. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            # 2. Decide what to do based on condition
-            if condition_found:
-                self.log("  - Condition met. Exiting loop.")
-                if step_id in self.loop_counters:
-                    del self.loop_counters[step_id]
-                self.execution_stack[-1] = (context['sequence'], context['index'] + 1)
-            else:
-                self.log("  - Condition not met. Performing loop actions.")
-                self.loop_counters[step_id] = current_retries + 1
-                # Push the loop step and then the sub-actions, just like in 'repeat' mode
-                self.execution_stack.append((context['sequence'], context['index']))
-                self.execution_stack.append((step.get('loop_actions', []), 0))
-
-            self.scan_job = self.after(10, self.run_scan_loop)
-
-    def _execute_single_action(self, action_step, scan_region):
-        """
-        Executes a single, simple action. Returns True on success, False on failure.
-        """
-        self.log(f"  - Action: Find {action_step['detection_mode']} '{action_step.get('detection_target_name', 'N/A')}'...")
-        haystack_img = capture_screen(scan_region)
-
-        target_pos = None
-        if action_step['detection_mode'] == "Color":
-            locations = find_color(haystack_img, action_step['detection_target'])
-            if locations:
-                target_pos = locations[0]
-        elif action_step['detection_mode'] == "Image":
-            try:
-                targets = action_step['detection_target']
-                if isinstance(targets, str): # Backward compatibility
-                    targets = [targets]
-
-                needle_imgs = []
-                for target_path in targets:
-                    img = cv2.imread(target_path, cv2.IMREAD_UNCHANGED)
-                    if img is None:
-                        self.log(f"    - Warning: Could not load template image '{os.path.basename(target_path)}'. Skipping it.")
-                        continue
-                    needle_imgs.append(img)
-
-                if not needle_imgs:
-                    self.log(f"    - Error: No valid template images could be loaded.")
-                    return False
-
-                threshold = self.settings_manager.get_setting('image_similarity_threshold')
-                target_pos = find_image(haystack_img, needle_imgs, threshold=threshold)
-            except Exception as e:
-                self.log(f"    - Error during image search: {e}")
-                return False
-
-        if target_pos:
-            self.log(f"    - Target found.")
-            abs_x = scan_region['left'] + target_pos[0]
-            abs_y = scan_region['top'] + target_pos[1]
-
-            # --- Action Preview ---
-            preview_duration = 0.5 # in seconds
-            action_type = action_step['action_type']
-            if not self.dry_run_var.get():
-                # TODO: Make this a setting
-                if action_type in ["Click", "Right-click", "Click with Offset"]:
-                    self.log(f"    - Previewing action at ({abs_x}, {abs_y}) for {preview_duration}s...")
-                    preview = ActionPreview(self, abs_x, abs_y, duration=int(preview_duration * 1000))
-                    self.wait_window(preview)
-
-
-            action_params = action_step.get('action_params', {})
-
-            if self.dry_run_var.get():
-                self.log(f"    - [DRY RUN] Would perform action: {action_type}")
-            else:
-                if action_type == "Click":
-                    self.log(f"    - Performing action: Click at ({abs_x}, {abs_y})")
-                    click_at(abs_x, abs_y)
-                elif action_type == "Right-click":
-                    self.log(f"    - Performing action: Right-click at ({abs_x}, {abs_y})")
-                    right_click_at(abs_x, abs_y)
-                elif action_type == "Click with Offset":
-                    offset_x = action_params.get('click_offset_x', 0)
-                    offset_y = action_params.get('click_offset_y', 0)
-                    click_x = abs_x + offset_x
-                    click_y = abs_y + offset_y
-                    self.log(f"    - Performing action: Click with offset at ({click_x}, {click_y})")
-                    click_at(click_x, click_y)
-                elif action_type == "Type":
-                    text_to_type = action_params.get('text', '')
-                    substituted_text = self._substitute_variables(text_to_type)
-                    self.log(f"    - Performing action: Type '{substituted_text}'")
-                    type_text(substituted_text)
-                elif action_type == "Key Combo":
-                    key_combo = action_params.get('key_combo', '')
-                    substituted_combo = self._substitute_variables(key_combo)
-                    self.log(f"    - Performing action: Press keys '{substituted_combo}'")
-                    press_key_combination(substituted_combo)
-                elif action_type == "Scroll":
-                    direction = action_params.get('scroll_direction', 'Down')
-                    amount = action_params.get('scroll_amount', 5)
-                    self.log(f"    - Performing action: Scroll {direction} by {amount}")
-                    scroll_wheel(direction.lower(), amount)
-
-            # Sub-actions have waits too
-            self._handle_post_action_wait(action_step)
-            return True
-        else:
-            self.log("    - Target not found.")
-            return False
-
-    def _execute_simple_step(self, step, context):
-        target_window_title = step.get("window_title")
-        if not target_window_title:
-            self.log(f"Error in {context['number_str']}: No target window specified. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        try:
-            target_windows = gw.getWindowsWithTitle(target_window_title)
-            if not target_windows:
-                self.log(f"{context['number_str']}: Window '{target_window_title}' not found. Re-scanning...")
-                self.scan_job = self.after(2000, self.run_scan_loop)
-                return
-            target_window = target_windows[0]
-            # Default scan region is the entire window
-            scan_region = {'top': target_window.top, 'left': target_window.left, 'width': target_window.width, 'height': target_window.height}
-
-            # If a specific search region is defined for the step, use it instead
-            if step.get('search_region'):
-                custom_region = step['search_region']
-                scan_region = {
-                    'top': target_window.top + custom_region['y'],
-                    'left': target_window.left + custom_region['x'],
-                    'width': custom_region['width'],
-                    'height': custom_region['height']
-                }
-        except Exception as e:
-            self.log(f"Error getting window details: {e}. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        if self._execute_single_action(step, scan_region):
-            # On success, advance the index and reset retry count for this step
-            sequence, index = self.execution_stack[-1]
-            self.execution_stack[-1] = (sequence, index + 1)
-            self.step_retry_counts[id(step)] = 0
-            self.log(f"SUCCESS: {context['number_str']} completed.")
-        else:
-            # On failure, handle according to the policy
-            on_failure = step.get('on_failure', {'policy': 'Stop'})
-            policy = on_failure.get('policy', 'Stop')
-            self.log(f"FAILURE: {context['number_str']} failed. Policy: {policy}.")
-
-            if policy == 'Stop':
-                self.log("Stopping bot due to step failure.")
-                self.toggle_bot()
-                return
-
-            elif policy == 'Skip':
-                self.log("Skipping to next step.")
-                sequence, index = self.execution_stack[-1]
-                self.execution_stack[-1] = (sequence, index + 1)
-                self.scan_job = self.after(10, self.run_scan_loop) # Proceed immediately
-
-            elif policy == 'Retry':
-                max_retries = on_failure.get('retries', 3)
-                step_id = id(step)
-                current_retries = self.step_retry_counts.get(step_id, 0)
-
-                if current_retries < max_retries:
-                    self.step_retry_counts[step_id] = current_retries + 1
-                    self.log(f"Retrying step... (Attempt {current_retries + 1}/{max_retries})")
-                    self.scan_job = self.after(2000, self.run_scan_loop) # Retry same step
-                else:
-                    self.log(f"Step failed after {max_retries} retries. Stopping bot.")
-                    self.toggle_bot()
-
-    def _execute_conditional_loop_step(self, step, context):
-        max_retries = step.get('max_retries', 5)
-        step_id = id(step)
-        current_retries = self.conditional_loop_retry_counts.get(step_id, 0)
-
-        if current_retries >= max_retries:
-            self.log(f"Loop failed after {max_retries} retries for {context['number_str']}. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        # The window title for a conditional step is stored in the step itself
-        target_window_title = step.get("window_title")
-        if not target_window_title:
-            self.log(f"Error in {context['number_str']} (Conditional): No target window specified. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        try:
-            target_windows = gw.getWindowsWithTitle(target_window_title)
-            if not target_windows:
-                self.log(f"{context['number_str']} (Conditional): Window '{target_window_title}' not found. Retrying in 2s...")
-                self.scan_job = self.after(2000, self.run_scan_loop)
-                return
-            target_window = target_windows[0]
-            # Default scan region is the entire window
-            scan_region = {'top': target_window.top, 'left': target_window.left, 'width': target_window.width, 'height': target_window.height}
-
-            # If a specific search region is defined for the step, use it instead
-            if step.get('search_region'):
-                custom_region = step['search_region']
-                # The custom region coords are relative to the window, so we add the window's top-left corner
-                scan_region = {
-                    'top': target_window.top + custom_region['y'],
-                    'left': target_window.left + custom_region['x'],
-                    'width': custom_region['width'],
-                    'height': custom_region['height']
-                }
-                self.log(f"Using custom scan region for conditional loop: {scan_region}")
-
-        except Exception as e:
-            self.log(f"Error getting window details: {e}. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        # 1. Look for the primary target
-        primary_target = step.get('primary_target', {})
-        self.log(f"{context['number_str']} (Attempt {current_retries+1}/{max_retries}): Finding '{primary_target.get('detection_target_name')}'...")
-        haystack_img = capture_screen(scan_region)
-
-        primary_pos = None
-        try:
-            targets = primary_target['detection_target']
-            if isinstance(targets, str): targets = [targets]
-            needle_imgs = [cv2.imread(p, cv2.IMREAD_UNCHANGED) for p in targets]
-            threshold = self.settings_manager.get_setting('image_similarity_threshold')
-            primary_pos = find_image(haystack_img, [img for img in needle_imgs if img is not None], threshold=threshold)
-        except Exception as e:
-            self.log(f"Loop Error: Could not load primary target image(s). {e}")
-            self.toggle_bot()
-            return
-
-        if primary_pos:
-            # 2. If found, success! Move to next step.
-            self.log("Primary target found! Proceeding to next step.")
-            self.conditional_loop_retry_counts[step_id] = 0 # Reset counter for this specific step
-            sequence, index = self.execution_stack[-1]
-            self.execution_stack[-1] = (sequence, index + 1)
-            self._handle_post_action_wait(step)
-        else:
-            # 3. If not found, perform fallback action.
-            self.log("Primary target not found. Performing fallback action.")
-            self.conditional_loop_retry_counts[step_id] = current_retries + 1 # Increment counter
-
-            fallback_action = step.get('on_fail', {})
-            if not fallback_action:
-                self.log("No fallback action defined. Stopping bot.")
-                self.toggle_bot()
-                return
-
-            if self._perform_fallback_action(fallback_action, scan_region):
-                self.log("Fallback action successful. Retrying primary target in 2 seconds...")
-                self.scan_job = self.after(2000, self.run_scan_loop) # Re-run the same conditional step
-            else:
-                self.log("Fallback action failed. Stopping bot.")
-                self.toggle_bot()
-
-    def _perform_fallback_action(self, action_details, scan_region):
-        action_type = action_details.get('action_type')
-
-
-        if action_type == "Do Nothing":
-            self.log("Fallback action: Doing nothing.")
-            return True
-        elif action_type == "Click and Drag":
-            if self.dry_run_var.get():
-                self.log("Fallback action: [DRY RUN] Would perform 'Click and Drag'.")
-            else:
-                self.log("Fallback action: Performing 'Click and Drag'.")
-                params = action_details.get('action_params', {})
-                offset_x = params.get('drag_offset_x', 0)
-                offset_y = params.get('drag_offset_y', 0)
-
-                start_x = scan_region['left'] + scan_region['width'] // 2
-                start_y = scan_region['top'] + scan_region['height'] // 2
-                end_x = start_x + offset_x
-                end_y = start_y + offset_y
-
-                self.log(f"Dragging from window center ({start_x}, {start_y}) to ({end_x}, {end_y})")
-                click_and_drag(start_x, start_y, end_x, end_y)
-            return True
-
-        elif action_type == "Scroll":
-            if self.dry_run_var.get():
-                self.log("Fallback action: [DRY RUN] Would scroll.")
-            else:
-                params = action_details.get('action_params', {})
-                direction = params.get('scroll_direction', 'Down')
-                amount = params.get('scroll_amount', 5)
-                self.log(f"Fallback action: Scrolling {direction} by {amount}")
-                scroll_wheel(direction.lower(), amount)
-            return True
-        elif action_type == "Click" or action_type == "Click with Offset":
-            self.log(f"Fallback: Finding '{action_details.get('detection_target_name')}' for action '{action_type}'.")
-            haystack_img = capture_screen(scan_region)
-            target_pos = None
-            try:
-                targets = action_details['detection_target']
-                if isinstance(targets, str): targets = [targets]
-                needle_imgs = [cv2.imread(p, cv2.IMREAD_UNCHANGED) for p in targets]
-                threshold = self.settings_manager.get_setting('image_similarity_threshold')
-                target_pos = find_image(haystack_img, [img for img in needle_imgs if img is not None], threshold=threshold)
-            except Exception as e:
-                self.log(f"Fallback Error: Could not load image(s). {e}")
-                return False
-
-            if target_pos:
-                abs_x = scan_region['left'] + target_pos[0]
-                abs_y = scan_region['top'] + target_pos[1]
-
-                if self.dry_run_var.get():
-                    self.log(f"Fallback action: [DRY RUN] Would perform '{action_type}' at ({abs_x}, {abs_y})")
-                else:
-                    if action_type == "Click":
-                        self.log(f"Fallback action: Clicking at ({abs_x}, {abs_y})")
-                        click_at(abs_x, abs_y)
-                    else: # Click with Offset
-                        params = action_details.get('action_params', {})
-                        offset_x = params.get('click_offset_x', 0)
-                        offset_y = params.get('click_offset_y', 0)
-                        click_x = abs_x + offset_x
-                        click_y = abs_y + offset_y
-                        self.log(f"Fallback action: Click with offset at ({click_x}, {click_y})")
-                        click_at(click_x, click_y)
-                return True
-            else:
-                self.log("Fallback target not found.")
-                return False
-
-        self.log(f"Unknown fallback action type: {action_type}")
-        return False
-
-    def _execute_time_based_condition_step(self, step, context):
-        time_cond = step.get('time_condition', {})
-        hour = time_cond.get('hour')
-        minute = time_cond.get('minute')
-
-        if hour is None or minute is None:
-            self.log(f"Error in {context['number_str']}: Invalid time condition. Stopping bot.")
-            self.toggle_bot()
-            return
-
-        now = time.localtime()
-        time_key = (hour, minute)
-
-        if now.tm_hour == hour and now.tm_min == minute:
-            if time_key in self.time_condition_executed:
-                self.log(f"Time condition {hour:02d}:{minute:02d} already executed this cycle. Advancing.")
-                # Condition was met and executed, so we just advance the parent sequence
-                self.execution_stack[-1] = (context['sequence'], context['index'] + 1)
-                self.scan_job = self.after(10, self.run_scan_loop)
-                return
-
-            self.log(f"Time condition met: {hour:02d}:{minute:02d}. Executing actions.")
-            self.time_condition_executed.add(time_key) # Mark as executed
-
-            # Advance the parent sequence *before* pushing a new one.
-            self.execution_stack[-1] = (context['sequence'], context['index'] + 1)
-            actions = step.get('actions', [])
-            if actions:
-                self.execution_stack.append((actions, 0))
-            # After executing, immediately continue to the next step in the main loop
-            self.scan_job = self.after(10, self.run_scan_loop)
-        else:
-            self.log(f"Waiting for time condition: {hour:02d}:{minute:02d}. Current time: {now.tm_hour:02d}:{now.tm_min:02d}. Re-checking in 5 seconds.")
-            self.scan_job = self.after(5000, self.run_scan_loop)
-
-    def log(self, message):
-        timestamp = time.strftime("%H:%M:%S")
-        self.log_area.insert(tk.END, f"[{timestamp}] {message}\n")
-        self.log_area.see(tk.END)
-
-    def _bgr_to_hex(self, bgr_color):
-        b, g, r = bgr_color
-        return f"#{r:02x}{g:02x}{b:02x}".upper()
-
-class WindowSelector(tk.Toplevel):
-    def __init__(self, master, is_splash=False):
-        super().__init__(master)
-        self.master = master
-        self.is_splash = is_splash
-        self.title("Select Target Window")
-        self.geometry("350x450")
-
-        # Find the root App instance for theming
-        app = self.master
-        while not isinstance(app, App):
-            app = app.master
-
-        self.configure(bg=app.bg_color)
-        self.transient(master)
-        self.grab_set()
-
-        self.listbox = tk.Listbox(self, bg=app.widget_bg_color, fg=app.text_color, relief=tk.FLAT)
-        self.listbox.pack(pady=10, padx=10, fill="both", expand=True)
-
-        button_frame = ttk.Frame(self)
-        button_frame.pack(pady=10)
-
-        ttk.Button(button_frame, text="Refresh", command=self.populate_windows).pack(side="left", padx=5)
-        ttk.Button(button_frame, text="Select", command=self.on_select, style="Accent.TButton").pack(side="left", padx=5)
-
-        cancel_text = "Quit" if self.is_splash and isinstance(self.master, App) else "Cancel"
-        ttk.Button(button_frame, text=cancel_text, command=self.on_cancel).pack(side="left", padx=5)
-
-        self.populate_windows()
-        self.protocol("WM_DELETE_WINDOW", self.on_cancel)
-
-    def populate_windows(self):
-        self.listbox.delete(0, tk.END)
-        titles = sorted([title for title in gw.getAllTitles() if title])
-        for title in titles:
-            self.listbox.insert(tk.END, title)
-
-    def on_select(self):
-        selected_indices = self.listbox.curselection()
-        if not selected_indices:
-            return
-        selected_title = self.listbox.get(selected_indices[0])
-        if hasattr(self.master, 'on_window_selected'):
-            self.master.on_window_selected(selected_title)
-        self.destroy()
-
-    def on_cancel(self):
-        # If this is the initial selection dialog and no selection has been made, closing it quits the app.
-        if self.is_splash and isinstance(self.master, App) and not self.master.target_window_title.get():
-            self.master.destroy()
-        else:
-            self.destroy()
-
-
-class ScreenshotTaker(tk.Toplevel):
-    def __init__(self, master):
-        super().__init__(master)
-        self.master = master
-        self.start_x = None
-        self.start_y = None
-        self.rect = None
-
-        # Configure the Toplevel window
-        self.attributes('-fullscreen', True)
-        self.attributes('-alpha', 0.3)
-        self.attributes('-topmost', True)
-        self.transient(master)
-        self.grab_set()
-
-        self.canvas = tk.Canvas(self, cursor="cross")
-        self.canvas.pack(fill="both", expand=True)
-
-        self.canvas.bind("<ButtonPress-1>", self.on_button_press)
-        self.canvas.bind("<B1-Motion>", self.on_mouse_drag)
-        self.canvas.bind("<ButtonRelease-1>", self.on_button_release)
-        self.canvas.bind("<Escape>", lambda e: self.destroy())
-
-    def on_button_press(self, event):
-        self.start_x = self.canvas.canvasx(event.x)
-        self.start_y = self.canvas.canvasy(event.y)
-        # Create rectangle if it doesn't exist
-        if not self.rect:
-            self.rect = self.canvas.create_rectangle(self.start_x, self.start_y, self.start_x, self.start_y, outline='red', width=2)
-
-    def on_mouse_drag(self, event):
-        cur_x = self.canvas.canvasx(event.x)
-        cur_y = self.canvas.canvasy(event.y)
-        self.canvas.coords(self.rect, self.start_x, self.start_y, cur_x, cur_y)
-
-    def on_button_release(self, event):
-        end_x = self.canvas.canvasx(event.x)
-        end_y = self.canvas.canvasy(event.y)
-
-        # Check for a minimal size to avoid accidental clicks
-        if abs(end_x - self.start_x) < 5 or abs(end_y - self.start_y) < 5:
-            self.destroy()
-            return
-
-        # Determine the bounding box
-        x1 = int(min(self.start_x, end_x))
-        y1 = int(min(self.start_y, end_y))
-        x2 = int(max(self.start_x, end_x))
-        y2 = int(max(self.start_y, end_y))
-
-        # Hide this transparent window
-        self.withdraw()
-        # Give the window manager time to process the withdraw command
-        self.update_idletasks()
-        time.sleep(0.2)
-
-        # Capture the screen region
-        with mss.mss() as sct:
-            monitor = {"top": y1, "left": x1, "width": x2 - x1, "height": y2 - y1}
-            img = sct.grab(monitor)
-
-            # Convert to numpy array in BGR format for OpenCV
-            img_np = np.array(img)
-            img_bgr = cv2.cvtColor(img_np, cv2.COLOR_BGRA2BGR)
-
-        # Call the master's callback
-        self.master.on_screenshot_taken(img_bgr)
-
-        # Destroy this window
-        self.destroy()
-
-
-class RegionSelector(tk.Toplevel):
-    def __init__(self, master, callback):
-        super().__init__(master)
-        self.master = master
-        self.callback = callback
-        self.start_x = None
-        self.start_y = None
-        self.rect = None
-
-        self.attributes('-fullscreen', True)
-        self.attributes('-alpha', 0.3)
-        self.attributes('-topmost', True)
-        self.transient(master)
-        self.grab_set()
-
-        self.canvas = tk.Canvas(self, cursor="cross")
-        self.canvas.pack(fill="both", expand=True)
-
-        self.canvas.bind("<ButtonPress-1>", self.on_button_press)
-        self.canvas.bind("<B1-Motion>", self.on_mouse_drag)
-        self.canvas.bind("<ButtonRelease-1>", self.on_button_release)
-        self.canvas.bind("<Escape>", lambda e: self.destroy())
-
-    def on_button_press(self, event):
-        self.start_x = self.canvas.canvasx(event.x)
-        self.start_y = self.canvas.canvasy(event.y)
-        if not self.rect:
-            self.rect = self.canvas.create_rectangle(self.start_x, self.start_y, self.start_x, self.start_y, outline='green', width=2)
-
-    def on_mouse_drag(self, event):
-        cur_x = self.canvas.canvasx(event.x)
-        cur_y = self.canvas.canvasy(event.y)
-        self.canvas.coords(self.rect, self.start_x, self.start_y, cur_x, cur_y)
-
-    def on_button_release(self, event):
-        end_x = self.canvas.canvasx(event.x)
-        end_y = self.canvas.canvasy(event.y)
-
-        if abs(end_x - self.start_x) < 5 or abs(end_y - self.start_y) < 5:
-            self.destroy()
-            return
-
-        x = int(min(self.start_x, end_x))
-        y = int(min(self.start_y, end_y))
-        width = int(abs(end_x - self.start_x))
-        height = int(abs(end_y - self.start_y))
-
-        region = {'x': x, 'y': y, 'width': width, 'height': height}
-        self.callback(region)
-        self.destroy()
-
-
-class ColorSampler(tk.Toplevel):
-    def __init__(self, master):
-        super().__init__(master)
-        self.master = master
-        app = self.master.master # The App instance
-
-        # Hide the main window to get a clean screenshot
-        app.withdraw()
-        time.sleep(0.3) # Give WM time to hide the window
-
-        # Configure the sampler Toplevel window
-        self.attributes('-fullscreen', True)
-        self.attributes('-topmost', True)
-        self.transient(master)
-        self.grab_set()
-
-        # Take a screenshot of the primary monitor
-        with mss.mss() as sct:
-            # sct.monitors[0] is the entire virtual screen, [1] is the primary monitor
-            sct_img = sct.grab(sct.monitors[1])
-
-        # Convert to a PIL Image. mss provides .rgb for direct PIL compatibility.
-        self.pil_img = Image.frombytes("RGB", sct_img.size, sct_img.rgb)
-
-        # Now that the screenshot is taken, we can show the main window again
-        app.deiconify()
-
-        # Create a PhotoImage to display in Tkinter
-        self.tk_img = ImageTk.PhotoImage(self.pil_img)
-
-        self.canvas = tk.Canvas(self, cursor="tcross")
-        self.canvas.pack(fill="both", expand=True)
-        self.canvas.create_image(0, 0, image=self.tk_img, anchor="nw")
-
-        self.canvas.bind("<ButtonPress-1>", self.on_click)
-        self.canvas.bind("<Escape>", lambda e: self.destroy())
-
-    def on_click(self, event):
-        x, y = event.x, event.y
-        # Get the color from the original PIL image
-        rgb_color = self.pil_img.getpixel((x, y))
-        # Convert from tkinter's RGB to OpenCV's BGR
-        bgr_color = [rgb_color[2], rgb_color[1], rgb_color[0]]
-        # Pass the color to the parent's callback
-        self.master.on_color_sampled(bgr_color)
-        # Close the sampler
-        self.destroy()
-
-
-class StepEditor(tk.Toplevel):
-    def __init__(self, master, step_data=None, index=None, target_sequence_list=None, on_save_callback=None, is_sub_editor=False):
-        super().__init__(master)
-        self.master = master # This is the App instance
-        self.step_data = step_data if step_data else {}
-        self.index = index
-        self.target_sequence_list = target_sequence_list
-        self.on_save_callback = on_save_callback
-        self.is_sub_editor = is_sub_editor
-
-        self.title("Action Editor" if is_sub_editor else "Step Editor")
-
-        # Find the root App instance to access theme colors correctly in sub-editors
-        app = self.master
-        while not isinstance(app, App):
-            app = app.master
-        self.app = app # Store it for later use if needed
-
-        self.configure(bg=app.bg_color)
-        self.transient(self.master)
-
-        # Make the window resizable
-        self.resizable(True, True)
-
-        self.grab_set()
-
-        # --- VARS ---
-        self.step_type = tk.StringVar(value=self.step_data.get('step_type', 'simple'))
-
-        # Vars for Simple Action
-        self.detection_mode = tk.StringVar(value=self.step_data.get('detection_mode', 'Image'))
-        self.action_type = tk.StringVar(value=self.step_data.get('action_type', 'Click'))
-        self.simple_click_offset_x = tk.StringVar(value=self.step_data.get('action_params', {}).get('click_offset_x', '0'))
-        self.simple_click_offset_y = tk.StringVar(value=self.step_data.get('action_params', {}).get('click_offset_y', '0'))
-        self.text_to_type = tk.StringVar(value=self.step_data.get('action_params', {}).get('text', ''))
-        self.key_combo_text = tk.StringVar(value=self.step_data.get('action_params', {}).get('key_combo', 'ctrl+c'))
-        self.variable_name = tk.StringVar(value=self.step_data.get('action_params', {}).get('variable_name', ''))
-        self.variable_value = tk.StringVar(value=self.step_data.get('action_params', {}).get('variable_value', ''))
-        self.modify_variable_name = tk.StringVar(value=self.step_data.get('action_params', {}).get('modify_variable_name', ''))
-        self.modify_variable_operation = tk.StringVar(value=self.step_data.get('action_params', {}).get('modify_variable_operation', 'add'))
-        self.modify_variable_value = tk.StringVar(value=self.step_data.get('action_params', {}).get('modify_variable_value', '1'))
-        self.output_variable_name = tk.StringVar(value=self.step_data.get('action_params', {}).get('output_variable_name', 'ocr_result'))
-        self.ocr_region = self.step_data.get('action_params', {}).get('ocr_region')
-        self.ocr_region_label_var = tk.StringVar(value=self._get_ocr_region_display_text())
-        self.scroll_direction = tk.StringVar(value=self.step_data.get('action_params', {}).get('scroll_direction', 'Down'))
-        self.scroll_amount = tk.StringVar(value=self.step_data.get('action_params', {}).get('scroll_amount', '5'))
-        self.target_window_title = tk.StringVar(value=self.step_data.get('window_title', self.master.target_window_title.get() or ''))
-
-        self.search_region = self.step_data.get('search_region') # This is now a direct attribute
-        self.search_region_label_var = tk.StringVar(value=self._get_region_display_text())
-        if self.step_data.get('detection_mode') == 'Color':
-            self.target_color_bgr = self.step_data.get('detection_target', [0,0,255])
-        else:
-            self.target_color_bgr = [0,0,255] # Default value
-        # This var is legacy, from before multi-image support. It's not actively used, but we'll initialize it safely.
-        detection_target = self.step_data.get('detection_target')
-        initial_template_name = ''
-        if self.step_data.get('detection_mode') == 'Image' and detection_target:
-            if isinstance(detection_target, list):
-                if detection_target: # Ensure list is not empty
-                    initial_template_name = os.path.basename(detection_target[0])
-            elif isinstance(detection_target, str):
-                initial_template_name = os.path.basename(detection_target)
-        self.template_var = tk.StringVar(value=initial_template_name)
-
-        # Vars for Conditional Loop
-        self.max_retries = tk.StringVar(value=self.step_data.get('max_retries', '5'))
-        self.fallback_action_type = tk.StringVar(value=self.step_data.get('on_fail', {}).get('action_type', 'Click'))
-        self.fallback_drag_offset_x = tk.StringVar(value=self.step_data.get('on_fail', {}).get('action_params', {}).get('drag_offset_x', '0'))
-        self.fallback_drag_offset_y = tk.StringVar(value=self.step_data.get('on_fail', {}).get('action_params', {}).get('drag_offset_y', '0'))
-
-        # Vars for Conditional Branch (If/Else)
-        condition = self.step_data.get('condition', {})
-        self.if_variable = tk.StringVar(value=condition.get('variable', ''))
-        self.if_operator = tk.StringVar(value=condition.get('operator', 'equals'))
-        self.if_value = tk.StringVar(value=condition.get('value', ''))
-        self.if_branch = self.step_data.get('if_branch', [])
-        self.else_branch = self.step_data.get('else_branch', [])
-
-        # Vars for Loop Step
-        self.loop_mode = tk.StringVar(value=self.step_data.get('loop_mode', 'repeat'))
-        self.loop_repeat_count = tk.StringVar(value=self.step_data.get('loop_repeat_count', '5'))
-        self.loop_actions = self.step_data.get('loop_actions', [])
-        self.loop_max_retries = tk.StringVar(value=self.step_data.get('max_retries', '10'))
-
-        # Vars for Time-based Condition
-        self.time_condition_hour = tk.StringVar(value=self.step_data.get('time_condition', {}).get('hour', '12'))
-        self.time_condition_minute = tk.StringVar(value=self.step_data.get('time_condition', {}).get('minute', '00'))
-        self.time_based_actions = self.step_data.get('actions', [])
-
-        # Vars for Post-Action Wait
-        wait_params = self.step_data.get('wait_params')
-        if wait_params is None: # It's a new step, so load the defaults from settings
-            default_wait_params = self.app.settings_manager.get_setting('default_wait_times')
-            self.wait_type = tk.StringVar(value='None') # Default to None for new steps
-            self.fixed_wait = tk.StringVar(value=str(default_wait_params.get('fixed_time', 1)))
-            self.min_wait = tk.StringVar(value=str(default_wait_params.get('min_time', 1)))
-            self.max_wait = tk.StringVar(value=str(default_wait_params.get('max_time', 2)))
-        else: # It's an existing step, load its saved values
-            self.wait_type = tk.StringVar(value=wait_params.get('type', 'None'))
-            self.fixed_wait = tk.StringVar(value=str(wait_params.get('fixed_time', 1)))
-            self.min_wait = tk.StringVar(value=str(wait_params.get('min_time', 1)))
-            self.max_wait = tk.StringVar(value=str(wait_params.get('max_time', 2)))
-
-        # Vars for On Failure
-        on_failure_params = self.step_data.get('on_failure', {})
-        self.on_failure_policy = tk.StringVar(value=on_failure_params.get('policy', 'Stop'))
-        self.on_failure_retries = tk.StringVar(value=on_failure_params.get('retries', '3'))
-
-
-        # --- LAYOUT FRAMES ---
-        # A bottom frame for buttons that never gets pushed out of view
-        button_frame = ttk.Frame(self)
-        button_frame.pack(side="bottom", fill="x", pady=10, padx=10)
-
-        # A main content frame that can expand and scroll
-        content_frame = ttk.Frame(self, padding="10")
-        content_frame.pack(side="top", fill="both", expand=True)
-        content_frame.columnconfigure(0, weight=1)
-
-        # --- WIDGETS ---
-        # --- Step Type Selection ---
-        step_type_frame = ttk.LabelFrame(content_frame, text="Step Type", padding="10")
-        step_type_frame.grid(row=0, column=0, sticky="ew", pady=(0, 10))
-        step_type_frame.columnconfigure(0, weight=1) # Allow radio buttons to space out
-
-        self.step_type_radios = {}
-        step_types = [("Simple Action", "simple"), ("If/Else", "conditional_branch"), ("Loop", "loop"), ("Time-based Condition", "time_based_condition"),("Conditional (Legacy)", "conditional_loop")]
-        for i, (text, value) in enumerate(step_types):
-            radio = ttk.Radiobutton(step_type_frame, text=text, variable=self.step_type, value=value, command=self.on_step_type_change)
-            radio.grid(row=0, column=i, sticky="ew", padx=5)
-            self.step_type_radios[value] = radio
-
-        # --- Main Frames for each step type (parented to content_frame) ---
-        self.simple_action_frame = ttk.Frame(content_frame, padding="10")
-        self.conditional_loop_frame = ttk.Frame(content_frame, padding="10")
-        self.loop_frame = ttk.Frame(content_frame, padding="10")
-        self.conditional_branch_frame = ttk.Frame(content_frame, padding="10")
-        self.time_based_condition_frame = ttk.Frame(content_frame, padding="10")
-
-        # --- UI for Simple Action Frame ---
-        self.build_simple_action_ui(self.simple_action_frame)
-
-        # --- UI for Conditional Loop Frame ---
-        self.build_conditional_loop_ui(self.conditional_loop_frame)
-
-        # --- UI for Loop Frame ---
-        self.build_loop_ui(self.loop_frame)
-
-        # --- UI for Conditional Branch Frame ---
-        self.build_conditional_branch_ui(self.conditional_branch_frame)
-
-        # --- UI for Time-based Condition Frame ---
-        self.build_time_based_condition_ui(self.time_based_condition_frame)
-
-        # --- Save/Cancel Buttons (parented to button_frame) ---
-        ttk.Button(button_frame, text="Cancel", command=self.destroy).pack(side="right", padx=10)
-        ttk.Button(button_frame, text="Save Step", command=self.on_save, style="Accent.TButton").pack(side="right")
-
-        self.on_step_type_change() # Set initial view
-
-    def build_simple_action_ui(self, parent_frame):
-        # This function builds the UI for a simple action, parented to the given frame.
-        parent_frame.columnconfigure(1, weight=1) # Allow labels and entries to expand
-
-        window_frame = ttk.LabelFrame(parent_frame, text="1. Select Target Window", padding="10")
-        window_frame.grid(row=0, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        window_frame.columnconfigure(0, weight=1)
-        self.window_label = ttk.Label(window_frame, textvariable=self.target_window_title, wraplength=250)
-        self.window_label.grid(row=0, column=0, sticky="ew", padx=(0, 10))
-        ttk.Button(window_frame, text="Select...", command=self.select_window).grid(row=0, column=1)
-        if not self.target_window_title.get(): self.target_window_title.set("(None Selected)")
-
-        region_frame = ttk.LabelFrame(parent_frame, text="2. Set Search Region (Optional)", padding="10")
-        region_frame.grid(row=1, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        region_frame.columnconfigure(0, weight=1)
-        self.region_label = ttk.Label(region_frame, textvariable=self.search_region_label_var, wraplength=350)
-        self.region_label.grid(row=0, column=0, sticky="ew", padx=(0, 10))
-        ttk.Button(region_frame, text="Set Region", command=self.set_search_region).grid(row=0, column=1, padx=(0, 5))
-        ttk.Button(region_frame, text="Clear", command=self.clear_search_region).grid(row=0, column=2)
-
-
-        mode_frame = ttk.LabelFrame(parent_frame, text="3. Choose What to Look For", padding="10")
-        mode_frame.grid(row=2, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        ttk.Radiobutton(mode_frame, text="Color", variable=self.detection_mode, value="Color", command=self.on_mode_change).pack(anchor="w")
-        ttk.Radiobutton(mode_frame, text="Image", variable=self.detection_mode, value="Image", command=self.on_mode_change).pack(anchor="w")
-
-        self.color_frame = ttk.Frame(parent_frame)
-        self.image_frame = ttk.Frame(parent_frame)
-
-        ttk.Button(self.color_frame, text="Sample Color", command=self.sample_color).pack()
-        self.color_preview = tk.Frame(self.color_frame, bg=self.app._bgr_to_hex(self.target_color_bgr), width=25, height=25, relief=tk.SUNKEN, borderwidth=1)
-        self.color_preview.pack(pady=5)
-
-        ttk.Button(self.image_frame, text="Take Screenshot", command=self.take_screenshot).pack(pady=(0,5))
-
-        # --- Image List ---
-        image_list_frame = ttk.Frame(self.image_frame)
-        image_list_frame.pack(fill="x", expand=True, pady=5)
-        image_list_frame.columnconfigure(0, weight=1)
-
-        self.image_listbox = tk.Listbox(image_list_frame, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=4, selectmode=tk.EXTENDED)
-        self.image_listbox.grid(row=0, column=0, sticky="ew", rowspan=2)
-
-        image_button_frame = ttk.Frame(image_list_frame)
-        image_button_frame.grid(row=0, column=1, rowspan=2, sticky="ns", padx=(5,0))
-        ttk.Button(image_button_frame, text="Add", command=lambda: self._add_image_template_to_listbox(self.image_listbox)).pack(fill="x", pady=2)
-        ttk.Button(image_button_frame, text="Remove", command=lambda: self._remove_image_template_from_listbox(self.image_listbox)).pack(fill="x", pady=2)
-
-        self._update_image_listbox()
-
-        action_frame = ttk.LabelFrame(parent_frame, text="4. Choose Action", padding="10")
-        action_frame.grid(row=4, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        action_types = ["Click", "Right-click", "Click with Offset", "Type", "Key Combo", "Scroll", "Set Variable", "Modify Variable", "OCR"]
-        for i, action in enumerate(action_types):
-            ttk.Radiobutton(action_frame, text=action, variable=self.action_type, value=action, command=self.on_action_change).grid(row=i, column=0, sticky="w")
-
-        self.type_entry_frame = ttk.Frame(action_frame)
-        self.type_entry = ttk.Entry(self.type_entry_frame, textvariable=self.text_to_type)
-        self.type_entry.pack(fill="x", padx=5, pady=5)
-
-        self.key_combo_frame = ttk.Frame(action_frame)
-        ttk.Label(self.key_combo_frame, text="Keys (e.g., ctrl+alt+delete):").pack(side="left", padx=5)
-        self.key_combo_entry = ttk.Entry(self.key_combo_frame, textvariable=self.key_combo_text, width=20)
-        self.key_combo_entry.pack(side="left", fill="x", expand=True, padx=5, pady=5)
-
-        self.set_variable_frame = ttk.Frame(action_frame)
-        ttk.Label(self.set_variable_frame, text="Name:").pack(side="left", padx=5)
-        ttk.Entry(self.set_variable_frame, textvariable=self.variable_name, width=15).pack(side="left", padx=5)
-        ttk.Label(self.set_variable_frame, text="Value:").pack(side="left", padx=5)
-        ttk.Entry(self.set_variable_frame, textvariable=self.variable_value, width=20).pack(side="left", padx=5)
-
-        self.modify_variable_frame = ttk.Frame(action_frame)
-        ttk.Label(self.modify_variable_frame, text="Name:").pack(side="left", padx=5)
-        ttk.Entry(self.modify_variable_frame, textvariable=self.modify_variable_name, width=15).pack(side="left", padx=5)
-        operations = ["add", "subtract", "set"]
-        ttk.OptionMenu(self.modify_variable_frame, self.modify_variable_operation, *operations).pack(side="left", padx=5)
-        ttk.Label(self.modify_variable_frame, text="Value:").pack(side="left", padx=5)
-        ttk.Entry(self.modify_variable_frame, textvariable=self.modify_variable_value, width=15).pack(side="left", padx=5)
-
-        self.ocr_frame = ttk.Frame(action_frame)
-        ocr_var_frame = ttk.Frame(self.ocr_frame)
-        ttk.Label(ocr_var_frame, text="Save Text to Variable:").pack(side="left", padx=5)
-        ttk.Entry(ocr_var_frame, textvariable=self.output_variable_name, width=20).pack(side="left", padx=5)
-        ocr_var_frame.pack(fill="x", pady=2)
-
-        ocr_region_frame = ttk.Frame(self.ocr_frame)
-        self.ocr_region_label = ttk.Label(ocr_region_frame, textvariable=self.ocr_region_label_var, wraplength=350)
-        self.ocr_region_label.pack(side="left", fill="x", expand=True, padx=5)
-        ttk.Button(ocr_region_frame, text="Set Region", command=self.set_ocr_region).pack(side="left", padx=(0, 5))
-        ocr_region_frame.pack(fill="x", pady=2)
-
-
-        self.simple_offset_frame = ttk.Frame(action_frame)
-        simple_offset_x_frame = ttk.Frame(self.simple_offset_frame)
-        ttk.Label(simple_offset_x_frame, text="X Offset:").pack(side="left", padx=5)
-        ttk.Entry(simple_offset_x_frame, textvariable=self.simple_click_offset_x, width=7).pack(side="left")
-        simple_offset_x_frame.pack(fill="x", pady=2)
-        simple_offset_y_frame = ttk.Frame(self.simple_offset_frame)
-        ttk.Label(simple_offset_y_frame, text="Y Offset:").pack(side="left", padx=5)
-        ttk.Entry(simple_offset_y_frame, textvariable=self.simple_click_offset_y, width=7).pack(side="left")
-        simple_offset_y_frame.pack(fill="x", pady=2)
-
-        self.scroll_frame = ttk.Frame(action_frame)
-        scroll_direction_frame = ttk.Frame(self.scroll_frame)
-        ttk.Label(scroll_direction_frame, text="Direction:").pack(side="left", padx=5)
-        ttk.OptionMenu(scroll_direction_frame, self.scroll_direction, "Down", "Up").pack(side="left")
-        scroll_direction_frame.pack(fill="x", pady=2)
-        scroll_amount_frame = ttk.Frame(self.scroll_frame)
-        ttk.Label(scroll_amount_frame, text="Amount:").pack(side="left", padx=5)
-        ttk.Entry(scroll_amount_frame, textvariable=self.scroll_amount, width=7).pack(side="left")
-        scroll_amount_frame.pack(fill="x", pady=2)
-
-        self.on_mode_change()
-        self.on_action_change()
-        self._build_on_failure_ui(parent_frame).grid(row=5, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        self._build_wait_ui(parent_frame).grid(row=6, column=0, columnspan=2, sticky="ew")
-
-    def _build_on_failure_ui(self, parent_frame):
-        on_failure_frame = ttk.LabelFrame(parent_frame, text="On Failure", padding="10")
-
-        # --- Policy ---
-        policy_frame = ttk.Frame(on_failure_frame)
-        ttk.Label(policy_frame, text="Action on Failure:").pack(side="left", padx=(0,10))
-        ttk.Radiobutton(policy_frame, text="Stop Bot", variable=self.on_failure_policy, value="Stop", command=self.on_failure_policy_change).pack(side="left")
-        ttk.Radiobutton(policy_frame, text="Skip Step", variable=self.on_failure_policy, value="Skip", command=self.on_failure_policy_change).pack(side="left")
-        ttk.Radiobutton(policy_frame, text="Retry Step", variable=self.on_failure_policy, value="Retry", command=self.on_failure_policy_change).pack(side="left")
-        policy_frame.pack(fill="x", pady=(0,5))
-
-        # --- Retries Frame ---
-        self.retries_frame = ttk.Frame(on_failure_frame)
-        ttk.Label(self.retries_frame, text="Number of Retries:").pack(side="left", padx=5)
-        self.retries_entry = ttk.Entry(self.retries_frame, textvariable=self.on_failure_retries, width=7)
-        self.retries_entry.pack(side="left")
-
-        self.on_failure_policy_change() # Set initial visibility
-        return on_failure_frame
-
-    def on_failure_policy_change(self):
-        if self.on_failure_policy.get() == "Retry":
-            self.retries_frame.pack(fill="x", pady=2)
-        else:
-            self.retries_frame.pack_forget()
-
-    def build_conditional_loop_ui(self, parent_frame):
-        # --- Loop Settings ---
-        retries_frame = ttk.LabelFrame(parent_frame, text="Loop Settings", padding="10")
-        retries_frame.pack(pady=5, padx=10, fill="x")
-        ttk.Label(retries_frame, text="Max Retries:").pack(side="left", padx=5)
-        ttk.Entry(retries_frame, textvariable=self.max_retries, width=5).pack(side="left", padx=5)
-
-        # --- Primary Target ---
-        primary_target_frame = ttk.LabelFrame(parent_frame, text="Primary Target (Image(s) to find)", padding="10")
-        primary_target_frame.pack(pady=5, padx=10, fill="x")
-
-        primary_list_frame = ttk.Frame(primary_target_frame)
-        primary_list_frame.pack(fill="x", expand=True, pady=5)
-
-        self.primary_image_listbox = tk.Listbox(primary_list_frame, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=4, selectmode=tk.EXTENDED)
-        self.primary_image_listbox.pack(side="left", fill="x", expand=True)
-
-        primary_button_frame = ttk.Frame(primary_list_frame)
-        primary_button_frame.pack(side="left", padx=(5,0))
-        ttk.Button(primary_button_frame, text="Add", command=lambda: self._add_image_template_to_listbox(self.primary_image_listbox)).pack(fill="x", pady=2)
-        ttk.Button(primary_button_frame, text="Remove", command=lambda: self._remove_image_template_from_listbox(self.primary_image_listbox)).pack(fill="x", pady=2)
-
-        # --- Fallback Action ---
-        fallback_action_frame = ttk.LabelFrame(parent_frame, text="Fallback Action (If primary target not found)", padding="10")
-        fallback_action_frame.pack(pady=5, padx=10, fill="x")
-
-        # --- Fallback Action Type ---
-        fallback_action_type_frame = ttk.Frame(fallback_action_frame)
-        ttk.Label(fallback_action_type_frame, text="Action:").pack(side="left", pady=2, padx=5)
-        ttk.Radiobutton(fallback_action_type_frame, text="Click", variable=self.fallback_action_type, value="Click", command=self.on_fallback_action_change).pack(side="left")
-        ttk.Radiobutton(fallback_action_type_frame, text="Click with Offset", variable=self.fallback_action_type, value="Click with Offset", command=self.on_fallback_action_change).pack(side="left")
-        ttk.Radiobutton(fallback_action_type_frame, text="Click and Drag", variable=self.fallback_action_type, value="Click and Drag", command=self.on_fallback_action_change).pack(side="left")
-        ttk.Radiobutton(fallback_action_type_frame, text="Do Nothing", variable=self.fallback_action_type, value="Do Nothing", command=self.on_fallback_action_change).pack(side="left")
-        fallback_action_type_frame.pack(fill="x")
-
-        # --- Fallback Action Params ---
-        self.fallback_drag_frame = ttk.Frame(fallback_action_frame)
-        drag_x_frame = ttk.Frame(self.fallback_drag_frame)
-        ttk.Label(drag_x_frame, text="X Offset:").pack(side="left", padx=5)
-        ttk.Entry(drag_x_frame, textvariable=self.fallback_drag_offset_x, width=7).pack(side="left")
-        drag_x_frame.pack(fill="x", pady=2)
-        drag_y_frame = ttk.Frame(self.fallback_drag_frame)
-        ttk.Label(drag_y_frame, text="Y Offset:").pack(side="left", padx=5)
-        ttk.Entry(drag_y_frame, textvariable=self.fallback_drag_offset_y, width=7).pack(side="left")
-        drag_y_frame.pack(fill="x", pady=2)
-
-        # --- Fallback Target ---
-        self.fallback_target_frame = ttk.Frame(fallback_action_frame)
-        ttk.Label(self.fallback_target_frame, text="Target for Fallback Action:").pack(pady=2, anchor="w", padx=5)
-
-        fallback_list_frame = ttk.Frame(self.fallback_target_frame)
-        fallback_list_frame.pack(fill="x", expand=True)
-
-        self.fallback_image_listbox = tk.Listbox(fallback_list_frame, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=4, selectmode=tk.EXTENDED)
-        self.fallback_image_listbox.pack(side="left", fill="x", expand=True)
-
-        fallback_button_frame = ttk.Frame(fallback_list_frame)
-        fallback_button_frame.pack(side="left", padx=(5,0))
-        ttk.Button(fallback_button_frame, text="Add", command=lambda: self._add_image_template_to_listbox(self.fallback_image_listbox)).pack(fill="x", pady=2)
-        ttk.Button(fallback_button_frame, text="Remove", command=lambda: self._remove_image_template_from_listbox(self.fallback_image_listbox)).pack(fill="x", pady=2)
-
-        self.fallback_target_frame.pack(fill="x")
-
-        self._update_primary_image_listbox()
-        self._update_fallback_image_listbox()
-        self.on_fallback_action_change()
-        self._build_wait_ui(parent_frame).pack(pady=10, padx=10, fill="x")
-
-    def _update_primary_image_listbox(self):
-        primary_target = self.step_data.get('primary_target', {})
-        self._populate_listbox_from_step_data(self.primary_image_listbox, primary_target, 'detection_target')
-
-    def build_loop_ui(self, parent_frame):
-        # --- Loop Mode ---
-        loop_mode_frame = ttk.LabelFrame(parent_frame, text="Loop Mode", padding="10")
-        loop_mode_frame.pack(pady=5, padx=10, fill="x")
-        ttk.Radiobutton(loop_mode_frame, text="Repeat X Times", variable=self.loop_mode, value="repeat", command=self.on_loop_mode_change).pack(side="left")
-        ttk.Radiobutton(loop_mode_frame, text="Until Condition Met", variable=self.loop_mode, value="until", command=self.on_loop_mode_change).pack(side="left")
-
-        # --- Loop Settings ---
-        self.loop_settings_frame = ttk.Frame(parent_frame)
-        self.loop_settings_frame.pack(pady=5, padx=10, fill="x")
-
-        self.repeat_frame = ttk.Frame(self.loop_settings_frame)
-        ttk.Label(self.repeat_frame, text="Repetitions:").pack(side="left", padx=5)
-        ttk.Entry(self.repeat_frame, textvariable=self.loop_repeat_count, width=5).pack(side="left")
-
-        self.until_frame = ttk.Frame(self.loop_settings_frame)
-        ttk.Label(self.until_frame, text="Condition (Image):").pack(anchor="w", padx=5)
-
-        until_list_frame = ttk.Frame(self.until_frame)
-        until_list_frame.pack(fill="x", expand=True)
-
-        self.until_image_listbox = tk.Listbox(until_list_frame, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=4, selectmode=tk.EXTENDED)
-        self.until_image_listbox.pack(side="left", fill="x", expand=True)
-
-        until_button_frame = ttk.Frame(until_list_frame)
-        until_button_frame.pack(side="left", padx=(5,0))
-        ttk.Button(until_button_frame, text="Add", command=lambda: self._add_image_template_to_listbox(self.until_image_listbox)).pack(fill="x", pady=2)
-        ttk.Button(until_button_frame, text="Remove", command=lambda: self._remove_image_template_from_listbox(self.until_image_listbox)).pack(fill="x", pady=2)
-
-        max_retries_frame = ttk.Frame(self.until_frame)
-        max_retries_frame.pack(fill="x", pady=2, side="bottom")
-        ttk.Label(max_retries_frame, text="Max Retries:").pack(side="left", padx=5)
-        ttk.Entry(max_retries_frame, textvariable=self.loop_max_retries, width=5).pack(side="left")
-
-        # --- Actions Frame ---
-        actions_frame = ttk.LabelFrame(parent_frame, text="Actions to Loop", padding="10")
-        actions_frame.pack(pady=5, padx=10, fill="both", expand=True)
-
-        list_container = ttk.Frame(actions_frame)
-        list_container.pack(fill="both", expand=True)
-
-        self.loop_actions_listbox = tk.Listbox(list_container, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=6)
-        self.loop_actions_listbox.pack(side="left", fill="both", expand=True)
-        self.loop_actions_listbox.bind("<<ListboxSelect>>", self.on_loop_action_select)
-
-        seq_button_frame = ttk.Frame(list_container)
-        seq_button_frame.pack(side="left", padx=(5,0), fill="y")
-
-        ttk.Button(seq_button_frame, text="Add", command=self._add_loop_action).pack(pady=2, fill="x")
-        self.edit_loop_action_button = ttk.Button(seq_button_frame, text="Edit", command=self._edit_loop_action, state=tk.DISABLED)
-        self.edit_loop_action_button.pack(pady=2, fill="x")
-        self.remove_loop_action_button = ttk.Button(seq_button_frame, text="Remove", command=self._remove_loop_action, state=tk.DISABLED)
-        self.remove_loop_action_button.pack(pady=2, fill="x")
-
-        # Disable loop/conditional step types if this is a sub-editor
-        if self.is_sub_editor:
-            self.step_type_radios['loop'].config(state=tk.DISABLED)
-            self.step_type_radios['conditional_loop'].config(state=tk.DISABLED)
-            self.step_type.set('simple') # Default to simple action for sub-steps
-
-
-        self.on_loop_mode_change()
-        self._update_loop_actions_listbox()
-        self._update_until_image_listbox()
-
-    def build_conditional_branch_ui(self, parent_frame):
-        # --- Condition Builder ---
-        condition_frame = ttk.LabelFrame(parent_frame, text="Condition", padding="10")
-        condition_frame.pack(pady=5, padx=10, fill="x")
-
-        ttk.Label(condition_frame, text="If variable").pack(side="left", padx=5)
-        ttk.Entry(condition_frame, textvariable=self.if_variable, width=15).pack(side="left", padx=5)
-
-        operators = ["equals", "not equals", "contains", "not contains", "is greater than", "is less than"]
-        ttk.OptionMenu(condition_frame, self.if_operator, *operators).pack(side="left", padx=5)
-
-        ttk.Label(condition_frame, text="value").pack(side="left", padx=5)
-        ttk.Entry(condition_frame, textvariable=self.if_value, width=15).pack(side="left", padx=5)
-
-        # --- Branch Frames ---
-        branch_parent_frame = ttk.Frame(parent_frame)
-        branch_parent_frame.pack(pady=5, padx=10, fill="both", expand=True)
-
-        # --- IF Branch ---
-        if_frame = self._create_branch_frame(branch_parent_frame, "IF Actions (if condition is true)")
-        self.if_listbox = self._create_branch_listbox(if_frame, self.on_if_action_select)
-        self.if_edit_button, self.if_remove_button = self._create_branch_buttons(
-            if_frame,
-            self._add_if_action,
-            self._edit_if_action,
-            self._remove_if_action
-        )
-        if_frame.pack(side="left", fill="both", expand=True, padx=(0, 5))
-
-        # --- ELSE Branch ---
-        else_frame = self._create_branch_frame(branch_parent_frame, "ELSE Actions (optional)")
-        self.else_listbox = self._create_branch_listbox(else_frame, self.on_else_action_select)
-        self.else_edit_button, self.else_remove_button = self._create_branch_buttons(
-            else_frame,
-            self._add_else_action,
-            self._edit_else_action,
-            self._remove_else_action
-        )
-        else_frame.pack(side="right", fill="both", expand=True, padx=(5, 0))
-
-        self._update_if_actions_listbox()
-        self._update_else_actions_listbox()
-
-
-    def _create_branch_frame(self, parent, title):
-        frame = ttk.LabelFrame(parent, text=title, padding="10")
-        return frame
-
-    def _create_branch_listbox(self, parent, select_callback):
-        list_container = ttk.Frame(parent)
-        list_container.pack(fill="both", expand=True)
-        listbox = tk.Listbox(list_container, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=8)
-        listbox.pack(side="left", fill="both", expand=True)
-        listbox.bind("<<ListboxSelect>>", select_callback)
-        return listbox
-
-    def _create_branch_buttons(self, parent, add_cmd, edit_cmd, remove_cmd):
-        button_frame = ttk.Frame(parent.winfo_children()[0]) # Get the list_container
-        button_frame.pack(side="left", padx=(5,0), fill="y")
-        ttk.Button(button_frame, text="Add", command=add_cmd).pack(pady=2, fill="x")
-        edit_button = ttk.Button(button_frame, text="Edit", command=edit_cmd, state=tk.DISABLED)
-        edit_button.pack(pady=2, fill="x")
-        remove_button = ttk.Button(button_frame, text="Remove", command=remove_cmd, state=tk.DISABLED)
-        remove_button.pack(pady=2, fill="x")
-        return edit_button, remove_button
-
-    def _add_if_action(self):
-        self._open_sub_editor(self.if_branch, self._update_if_actions_listbox)
-    def _edit_if_action(self):
-        self._open_sub_editor(self.if_branch, self._update_if_actions_listbox, self.if_listbox.curselection())
-    def _remove_if_action(self):
-        self._remove_action_from_branch(self.if_branch, self.if_listbox.curselection(), self._update_if_actions_listbox)
-
-    def _add_else_action(self):
-        self._open_sub_editor(self.else_branch, self._update_else_actions_listbox)
-    def _edit_else_action(self):
-        self._open_sub_editor(self.else_branch, self._update_else_actions_listbox, self.else_listbox.curselection())
-    def _remove_else_action(self):
-        self._remove_action_from_branch(self.else_branch, self.else_listbox.curselection(), self._update_else_actions_listbox)
-
-    def _open_sub_editor(self, branch_list, update_callback, selection=None):
-        index = None
-        step_data = {}
-        if selection:
-            index = selection[0]
-            step_data = branch_list[index]
-        StepEditor(
-            master=self,
-            step_data=step_data,
-            index=index,
-            target_sequence_list=branch_list,
-            on_save_callback=update_callback,
-            is_sub_editor=True
-        )
-
-    def _remove_action_from_branch(self, branch_list, selection, update_callback):
-        if not selection: return
-        index = selection[0]
-        branch_list.pop(index)
-        update_callback()
-        self.app.log(f"Removed sub-action {index+1}.")
-
-    def on_if_action_select(self, event):
-        self._update_button_state(self.if_listbox.curselection(), self.if_edit_button, self.if_remove_button)
-    def on_else_action_select(self, event):
-        self._update_button_state(self.else_listbox.curselection(), self.else_edit_button, self.else_remove_button)
-
-    def _update_button_state(self, selection, edit_button, remove_button):
-        if selection:
-            edit_button.config(state=tk.NORMAL)
-            remove_button.config(state=tk.NORMAL)
-        else:
-            edit_button.config(state=tk.DISABLED)
-            remove_button.config(state=tk.DISABLED)
-
-    def _update_if_actions_listbox(self):
-        self._update_branch_listbox(self.if_listbox, self.if_branch)
-    def _update_else_actions_listbox(self):
-        self._update_branch_listbox(self.else_listbox, self.else_branch)
-
-    def _update_branch_listbox(self, listbox, branch):
-        listbox.delete(0, tk.END)
-        for i, step in enumerate(branch):
-            action = step.get('action_type', '?')
-            target = step.get('detection_target_name', 'Unknown')
-            text = f"{i+1}: {action} on '{target}'"
-            listbox.insert(tk.END, text)
-        # Manually trigger select event to update buttons
-        self.on_if_action_select(None)
-        self.on_else_action_select(None)
-
-
-    def _add_time_based_action(self):
-        self._open_sub_editor(self.time_based_actions, self._update_time_based_actions_listbox)
-
-    def _edit_time_based_action(self):
-        self._open_sub_editor(self.time_based_actions, self._update_time_based_actions_listbox, self.time_based_actions_listbox.curselection())
-
-    def _remove_time_based_action(self):
-        self._remove_action_from_branch(self.time_based_actions, self.time_based_actions_listbox.curselection(), self._update_time_based_actions_listbox)
-
-    def _update_time_based_actions_listbox(self):
-        self._update_branch_listbox(self.time_based_actions_listbox, self.time_based_actions)
-        self.time_based_actions_listbox.bind("<<ListboxSelect>>", self.on_time_based_action_select)
-
-    def on_time_based_action_select(self, event):
-        self._update_button_state(self.time_based_actions_listbox.curselection(), self.edit_time_based_action_button, self.remove_time_based_action_button)
-
-    def _add_loop_action(self):
-        StepEditor(
-            master=self.master,
-            target_sequence_list=self.loop_actions,
-            on_save_callback=self._update_loop_actions_listbox,
-            is_sub_editor=True
-        )
-
-    def _edit_loop_action(self):
-        selected_indices = self.loop_actions_listbox.curselection()
-        if not selected_indices:
-            return
-        index = selected_indices[0]
-        step_data = self.loop_actions[index]
-
-        StepEditor(
-            master=self.master,
-            step_data=step_data,
-            index=index,
-            target_sequence_list=self.loop_actions,
-            on_save_callback=self._update_loop_actions_listbox,
-            is_sub_editor=True
-        )
-
-    def _remove_loop_action(self):
-        selected_indices = self.loop_actions_listbox.curselection()
-        if not selected_indices:
-            return
-        index = selected_indices[0]
-        self.loop_actions.pop(index)
-        self._update_loop_actions_listbox()
-        self.app.log(f"Removed sub-action {index+1}.")
-
-    def on_loop_action_select(self, event):
-        selected_indices = self.loop_actions_listbox.curselection()
-        if selected_indices:
-            self.edit_loop_action_button.config(state=tk.NORMAL)
-            self.remove_loop_action_button.config(state=tk.NORMAL)
-        else:
-            self.edit_loop_action_button.config(state=tk.DISABLED)
-            self.remove_loop_action_button.config(state=tk.DISABLED)
-
-    def _update_loop_actions_listbox(self):
-        self.loop_actions_listbox.delete(0, tk.END)
-        for i, step in enumerate(self.loop_actions):
-            # This is a simplified representation. We can enhance it later.
-            action = step.get('action_type', '?')
-            target = step.get('detection_target_name', 'Unknown')
-            text = f"{i+1}: {action} on '{target}'"
-            self.loop_actions_listbox.insert(tk.END, text)
-
-
-    def on_loop_mode_change(self):
-        mode = self.loop_mode.get()
-        if mode == 'repeat':
-            self.repeat_frame.pack(fill="x")
-            self.until_frame.pack_forget()
-        else: # until
-            self.repeat_frame.pack_forget()
-            self.until_frame.pack(fill="x")
-
-
-
-    def build_time_based_condition_ui(self, parent_frame):
-        # --- Time Settings ---
-        time_frame = ttk.LabelFrame(parent_frame, text="Time Condition", padding="10")
-        time_frame.pack(pady=5, padx=10, fill="x")
-        ttk.Label(time_frame, text="Hour (0-23):").pack(side="left", padx=5)
-        ttk.Entry(time_frame, textvariable=self.time_condition_hour, width=5).pack(side="left", padx=5)
-        ttk.Label(time_frame, text="Minute (0-59):").pack(side="left", padx=5)
-        ttk.Entry(time_frame, textvariable=self.time_condition_minute, width=5).pack(side="left", padx=5)
-
-        # Add a real-time clock label
-        self.current_time_label_var = tk.StringVar()
-        current_time_label = ttk.Label(time_frame, textvariable=self.current_time_label_var)
-        current_time_label.pack(side="left", padx=20)
-        self._update_clock()
-
-        actions_frame = ttk.LabelFrame(parent_frame, text="Actions to run at specified time", padding="10")
-        actions_frame.pack(pady=5, padx=10, fill="both", expand=True)
-
-        list_container = ttk.Frame(actions_frame)
-        list_container.pack(fill="both", expand=True)
-
-        self.time_based_actions_listbox = tk.Listbox(list_container, bg=self.app.widget_bg_color, fg=self.app.text_color, relief=tk.FLAT, height=6)
-        self.time_based_actions_listbox.pack(side="left", fill="both", expand=True)
-
-        seq_button_frame = ttk.Frame(list_container)
-        seq_button_frame.pack(side="left", padx=(5,0), fill="y")
-
-        ttk.Button(seq_button_frame, text="Add", command=self._add_time_based_action).pack(pady=2, fill="x")
-        self.edit_time_based_action_button = ttk.Button(seq_button_frame, text="Edit", command=self._edit_time_based_action, state=tk.DISABLED)
-        self.edit_time_based_action_button.pack(pady=2, fill="x")
-        self.remove_time_based_action_button = ttk.Button(seq_button_frame, text="Remove", command=self._remove_time_based_action, state=tk.DISABLED)
-        self.remove_time_based_action_button.pack(pady=2, fill="x")
-
-        self._update_time_based_actions_listbox()
-
-    def _update_clock(self):
-        now_str = time.strftime("%H:%M:%S")
-        self.current_time_label_var.set(f"Bot Time: {now_str}")
-        self._clock_job = self.after(1000, self._update_clock)
-
-    def destroy(self):
-        # Override destroy to cancel the clock job
-        if hasattr(self, '_clock_job'):
-            self.after_cancel(self._clock_job)
-        super().destroy()
-
-    def on_step_type_change(self):
-        step_type = self.step_type.get()
-        # Hide all frames first
-        self.simple_action_frame.grid_remove()
-        self.conditional_loop_frame.grid_remove()
-        self.loop_frame.grid_remove()
-        self.conditional_branch_frame.grid_remove()
-        self.time_based_condition_frame.grid_remove()
-
-        if step_type == 'simple':
-            self.simple_action_frame.grid(row=1, column=0, sticky="ew")
-        elif step_type == 'loop':
-            self.loop_frame.grid(row=1, column=0, sticky="ew")
-        elif step_type == 'conditional_branch':
-            self.conditional_branch_frame.grid(row=1, column=0, sticky="ew")
-        elif step_type == 'time_based_condition':
-            self.time_based_condition_frame.grid(row=1, column=0, sticky="ew")
-        else: # conditional_loop
-            self.conditional_loop_frame.grid(row=1, column=0, sticky="ew")
-
-    def on_mode_change(self):
-        if self.detection_mode.get() == "Color":
-            self.image_frame.grid_remove()
-            self.color_frame.grid(row=3, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-        else:
-            self.color_frame.grid_remove()
-            self.image_frame.grid(row=3, column=0, columnspan=2, sticky="ew", pady=(0, 10))
-
-    def on_action_change(self):
-        action = self.action_type.get()
-
-        # Hide all action-specific frames first
-        self.type_entry_frame.grid_remove()
-        self.simple_offset_frame.grid_remove()
-        self.scroll_frame.grid_remove()
-        self.key_combo_frame.grid_remove()
-        self.set_variable_frame.grid_remove()
-        self.modify_variable_frame.grid_remove()
-        self.ocr_frame.grid_remove()
-
-        # The grid row for these frames should be after the radio buttons
-        action_row = 10 # A number greater than the number of radio buttons
-
-        if action == "Type":
-            self.type_entry_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5)
-        elif action == "Click with Offset":
-            self.simple_offset_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5, padx=20)
-        elif action == "Key Combo":
-            self.key_combo_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5)
-        elif action == "Scroll":
-            self.scroll_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5, padx=20)
-        elif action == "Set Variable":
-            self.set_variable_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5)
-        elif action == "Modify Variable":
-            self.modify_variable_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5)
-        elif action == "OCR":
-            self.ocr_frame.grid(row=action_row, column=0, columnspan=2, sticky="ew", pady=5)
-
-    def on_fallback_action_change(self):
-        action = self.fallback_action_type.get()
-        if action == "Click with Offset":
-            self.fallback_drag_frame.pack(fill="x", padx=15, pady=2)
-            self.fallback_target_frame.pack(fill="x")
-        elif action == "Click and Drag":
-            self.fallback_drag_frame.pack(fill="x", padx=15, pady=2)
-            self.fallback_target_frame.pack_forget()
-        elif action == "Do Nothing":
-            self.fallback_drag_frame.pack_forget()
-            self.fallback_target_frame.pack_forget()
-        else: # Click
-            self.fallback_drag_frame.pack_forget()
-            self.fallback_target_frame.pack(fill="x")
-
-    def _build_wait_ui(self, parent_frame):
-        wait_frame = ttk.LabelFrame(parent_frame, text="Post-Action Wait", padding="10")
-
-        # --- Wait Type ---
-        wait_type_frame = ttk.Frame(wait_frame)
-        ttk.Radiobutton(wait_type_frame, text="None", variable=self.wait_type, value="None", command=self.on_wait_type_change).pack(side="left")
-        ttk.Radiobutton(wait_type_frame, text="Fixed", variable=self.wait_type, value="Fixed", command=self.on_wait_type_change).pack(side="left")
-        ttk.Radiobutton(wait_type_frame, text="Random", variable=self.wait_type, value="Random", command=self.on_wait_type_change).pack(side="left")
-        wait_type_frame.pack(fill="x", pady=(0,5))
-
-        # --- Fixed Wait Frame ---
-        self.fixed_wait_frame = ttk.Frame(wait_frame)
-        ttk.Label(self.fixed_wait_frame, text="Wait (sec):").pack(side="left", padx=5)
-        ttk.Entry(self.fixed_wait_frame, textvariable=self.fixed_wait, width=7).pack(side="left")
-
-        # --- Random Wait Frame ---
-        self.random_wait_frame = ttk.Frame(wait_frame)
-        ttk.Label(self.random_wait_frame, text="Min (sec):").pack(side="left", padx=5)
-        ttk.Entry(self.random_wait_frame, textvariable=self.min_wait, width=7).pack(side="left")
-        ttk.Label(self.random_wait_frame, text="Max (sec):").pack(side="left", padx=(10,5))
-        ttk.Entry(self.random_wait_frame, textvariable=self.max_wait, width=7).pack(side="left")
-
-        self.on_wait_type_change() # Set initial visibility
-        return wait_frame
-
-    def on_wait_type_change(self):
-        wait_type = self.wait_type.get()
-        if wait_type == "Fixed":
-            self.fixed_wait_frame.pack(fill="x", pady=2)
-            self.random_wait_frame.pack_forget()
-        elif wait_type == "Random":
-            self.fixed_wait_frame.pack_forget()
-            self.random_wait_frame.pack(fill="x", pady=2)
-        else: # None
-            self.fixed_wait_frame.pack_forget()
-            self.random_wait_frame.pack_forget()
-
-    def on_save(self):
-        step_type = self.step_type.get()
-
-        if step_type == 'simple':
-            step = {
-                "step_type": "simple",
-                "window_title": self.target_window_title.get(),
-                "detection_mode": self.detection_mode.get(),
-                "action_type": self.action_type.get(),
-                "action_params": {},
-                "detection_target": None,
-                "detection_target_name": "",
-                "search_region": self.search_region,
-                "on_failure": {}
-            }
-
-            # --- Save On Failure settings ---
-            on_failure_policy = self.on_failure_policy.get()
-            step['on_failure']['policy'] = on_failure_policy
-            if on_failure_policy == 'Retry':
-                try:
-                    step['on_failure']['retries'] = int(self.on_failure_retries.get())
-                except ValueError:
-                    self.app.log("Error: 'On Failure' retries must be an integer.")
-                    return
-
-            action_type = step['action_type']
-            if action_type == 'Type':
-                step['action_params']['text'] = self.text_to_type.get()
-            elif action_type == 'Key Combo':
-                step['action_params']['key_combo'] = self.key_combo_text.get()
-            elif action_type == 'Set Variable':
-                step['action_params']['variable_name'] = self.variable_name.get()
-                step['action_params']['variable_value'] = self.variable_value.get()
-            elif action_type == 'Modify Variable':
-                step['action_params']['modify_variable_name'] = self.modify_variable_name.get()
-                step['action_params']['modify_variable_operation'] = self.modify_variable_operation.get()
-                step['action_params']['modify_variable_value'] = self.modify_variable_value.get()
-            elif action_type == 'OCR':
-                step['action_params']['output_variable_name'] = self.output_variable_name.get()
-                step['action_params']['ocr_region'] = self.ocr_region
-                if not self.ocr_region:
-                    self.app.log("Error: OCR region is not set for this step.")
-                    return
-            elif action_type == 'Click with Offset':
-                try:
-                    step['action_params']['click_offset_x'] = int(self.simple_click_offset_x.get())
-                    step['action_params']['click_offset_y'] = int(self.simple_click_offset_y.get())
-                except ValueError:
-                    self.app.log("Error: Click offsets must be integers.")
-                    return
-            elif action_type == 'Scroll':
-                try:
-                    step['action_params']['scroll_direction'] = self.scroll_direction.get()
-                    step['action_params']['scroll_amount'] = int(self.scroll_amount.get())
-                except ValueError:
-                    self.app.log("Error: Scroll amount must be an integer.")
-                    return
-
-            if step['detection_mode'] == 'Color':
-                step['detection_target'] = self.target_color_bgr
-                step['detection_target_name'] = self.app._bgr_to_hex(self.target_color_bgr)
-            else: # Image
-                target_names = list(self.image_listbox.get(0, tk.END))
-                if not target_names:
-                    self.app.log("Error: No template images selected for this step.")
-                    return
-                # Save the list of full paths for later execution
-                step['detection_target'] = [os.path.join("templates", name) for name in target_names]
-                # Save just the names for display purposes
-                step['detection_target_name'] = ", ".join(target_names)
-
-        elif step_type == 'loop':
-            try:
-                repeat_count = int(self.loop_repeat_count.get())
-            except ValueError:
-                self.app.log("Error: Repetitions must be an integer.")
-                return
-
-            step = {
-                "step_type": "loop",
-                "loop_mode": self.loop_mode.get(),
-                "loop_actions": self.loop_actions,
-                "window_title": self.target_window_title.get(),
-                "search_region": self.search_region
-            }
-            if step['loop_mode'] == 'repeat':
-                step['loop_repeat_count'] = repeat_count
-            else: # until
-                try:
-                    max_retries = int(self.loop_max_retries.get())
-                except ValueError:
-                    self.app.log("Error: Max retries must be an integer.")
-                    return
-
-                condition_target_names = list(self.until_image_listbox.get(0, tk.END))
-                if not condition_target_names:
-                    self.app.log("Error: At least one condition target image must be selected for an 'until' loop.")
-                    return
-
-                step['loop_condition_target'] = [os.path.join("templates", name) for name in condition_target_names]
-                step['loop_condition_target_name'] = ", ".join(condition_target_names)
-                step['max_retries'] = max_retries
-
-        elif step_type == 'conditional_branch':
-            step = {
-                "step_type": "conditional_branch",
-                "condition": {
-                    "variable": self.if_variable.get(),
-                    "operator": self.if_operator.get(),
-                    "value": self.if_value.get()
-                },
-                "if_branch": self.if_branch,
-                "else_branch": self.else_branch,
-                "window_title": self.target_window_title.get(), # Retain for consistency, though not used by parent
-            }
-
-        elif step_type == 'time_based_condition':
-            try:
-                hour = int(self.time_condition_hour.get())
-                minute = int(self.time_condition_minute.get())
-                if not (0 <= hour <= 23 and 0 <= minute <= 59):
-                    raise ValueError("Hour must be between 0-23 and minute between 0-59.")
-            except ValueError as e:
-                self.app.log(f"Error: Invalid time condition. {e}")
-                return
-
-            step = {
-                "step_type": "time_based_condition",
-                "time_condition": {"hour": hour, "minute": minute},
-                "actions": self.time_based_actions,
-                "window_title": self.target_window_title.get(),
-            }
-        else: # conditional_loop
-            try:
-                max_retries = int(self.max_retries.get())
-            except ValueError:
-                self.app.log("Error: Max retries must be an integer.")
-                return
-
-            # Primary Target
-            primary_target_names = list(self.primary_image_listbox.get(0, tk.END))
-            if not primary_target_names:
-                self.app.log("Error: At least one primary target image must be selected.")
-                return
-
-            primary_target_dict = {
-                "detection_mode": "Image",
-                "detection_target": [os.path.join("templates", name) for name in primary_target_names],
-                "detection_target_name": ", ".join(primary_target_names),
-            }
-
-            # Fallback (on_fail) action
-            fallback_action_type = self.fallback_action_type.get()
-            on_fail_dict = { "action_type": fallback_action_type, "action_params": {} }
-
-            if fallback_action_type in ["Click", "Click with Offset"]:
-                fallback_target_names = list(self.fallback_image_listbox.get(0, tk.END))
-                if not fallback_target_names:
-                    self.app.log(f"Error: At least one fallback target image must be selected for a '{fallback_action_type}' action.")
-                    return
-                on_fail_dict["detection_mode"] = "Image"
-                on_fail_dict["detection_target"] = [os.path.join("templates", name) for name in fallback_target_names]
-                on_fail_dict["detection_target_name"] = ", ".join(fallback_target_names)
-
-            if fallback_action_type == "Click with Offset":
-                try:
-                    on_fail_dict['action_params']['click_offset_x'] = int(self.fallback_drag_offset_x.get())
-                    on_fail_dict['action_params']['click_offset_y'] = int(self.fallback_drag_offset_y.get())
-                except ValueError:
-                    self.app.log("Error: Fallback click offsets must be integers.")
-                    return
-            elif fallback_action_type == "Click and Drag":
-                try:
-                    on_fail_dict['action_params']['drag_offset_x'] = int(self.fallback_drag_offset_x.get())
-                    on_fail_dict['action_params']['drag_offset_y'] = int(self.fallback_drag_offset_y.get())
-                except ValueError:
-                    self.app.log("Error: Fallback drag offsets must be integers.")
-                    return
-
-            step = {
-                "step_type": "conditional_loop",
-                "window_title": self.target_window_title.get(),
-                "max_retries": max_retries,
-                "primary_target": primary_target_dict,
-                "on_fail": on_fail_dict,
-                "search_region": self.search_region
-            }
-
-        # --- Save Wait Parameters ---
-        wait_type = self.wait_type.get()
-        wait_params = {'type': wait_type}
-        try:
-            if wait_type == 'Fixed':
-                wait_params['fixed_time'] = float(self.fixed_wait.get())
-            elif wait_type == 'Random':
-                wait_params['min_time'] = float(self.min_wait.get())
-                wait_params['max_time'] = float(self.max_wait.get())
-                if wait_params['min_time'] >= wait_params['max_time']:
-                    self.app.log("Error: Min wait time must be less than max wait time.")
-                    return
-        except ValueError:
-            self.app.log("Error: Wait times must be valid numbers.")
-            return
-        step['wait_params'] = wait_params
-
-        if self.target_sequence_list is not None and self.on_save_callback is not None:
-            if self.index is not None:
-                self.target_sequence_list[self.index] = step
-            else:
-                self.target_sequence_list.append(step)
-            self.on_save_callback()
-        else:
-            # Default behavior: save to the main sequence
-            self.master.on_step_saved(step, self.index)
-
-        self.destroy()
-
-    def select_window(self):
-        WindowSelector(self, is_splash=False)
-
-    def on_window_selected(self, title):
-        self.target_window_title.set(title)
-        self.app.log(f"Step editor window target set to: {title}")
-
-    def sample_color(self):
-        sampler = ColorSampler(self)
-        self.wait_window(sampler)
-        self.grab_set() # Re-grab focus
-
-    def on_color_sampled(self, bgr_color):
-        self.target_color_bgr = bgr_color
-        hex_color = self.app._bgr_to_hex(bgr_color)
-        self.color_preview.config(bg=hex_color)
-        self.app.log(f"Step color changed to {hex_color}")
-
-    def take_screenshot(self):
-        taker = ScreenshotTaker(self)
-        self.wait_window(taker)
-        self.grab_set() # Re-grab focus
-
-    def on_screenshot_taken(self, image):
-        self.app.log("Screenshot captured for step.")
-        filepath = filedialog.asksaveasfilename(
-            parent=self,
-            initialdir="templates",
-            title="Save Template",
-            filetypes=(("PNG files", "*.png"),),
-            defaultextension=".png"
-        )
-        if filepath:
-            try:
-                cv2.imwrite(filepath, image)
-                self.app.log(f"Template saved to {filepath}")
-
-                # This method is now only called from the simple action editor,
-                # so we can assume it's for the main image listbox.
-                self.image_listbox.insert(tk.END, os.path.basename(filepath))
-
-            except Exception as e:
-                self.app.log(f"Error saving template: {e}")
-
-    def _get_region_display_text(self):
-        region = self.search_region
-        if region:
-            return f"Region set: X={region['x']}, Y={region['y']}, W={region['width']}, H={region['height']}"
-        return "Not set. The entire target window will be searched."
-
-    def set_search_region(self):
-        self.app.log("Opening region selector...")
-        selector = RegionSelector(self, self.on_region_selected)
-        self.wait_window(selector)
-        self.grab_set()
-
-    def on_region_selected(self, region):
-        # The region coordinates are relative to the screen. We need to make them
-        # relative to the target window if a window is selected.
-        try:
-            target_window_title = self.target_window_title.get()
-            if target_window_title and target_window_title != "(None Selected)":
-                target_windows = gw.getWindowsWithTitle(target_window_title)
-                if target_windows:
-                    win = target_windows[0]
-                    region['x'] -= win.left
-                    region['y'] -= win.top
-                    self.app.log(f"Region coordinates adjusted to be relative to window '{win.title}'.")
-
-        except Exception as e:
-            self.app.log(f"Could not adjust region to window: {e}. Using absolute coordinates.")
-
-        self.search_region = region
-        self.search_region_label_var.set(self._get_region_display_text())
-        self.app.log(f"Search region set.")
-
-    def clear_search_region(self):
-        self.search_region = None
-        self.search_region_label_var.set(self._get_region_display_text())
-        self.app.log("Search region has been cleared.")
-
-    def _add_image_template_to_listbox(self, listbox):
-        filepaths = filedialog.askopenfilenames(
-            parent=self,
-            initialdir="templates",
-            title="Select Image Templates",
-            filetypes=(("PNG files", "*.png"), ("All files", "*.*"))
-        )
-        if filepaths:
-            for filepath in filepaths:
-                filename = os.path.basename(filepath)
-                # Avoid adding duplicates
-                if filename not in listbox.get(0, tk.END):
-                    listbox.insert(tk.END, filename)
-
-    def _remove_image_template_from_listbox(self, listbox):
-        selected_indices = listbox.curselection()
-        # Reverse the indices to avoid issues when deleting multiple items
-        for i in sorted(selected_indices, reverse=True):
-            listbox.delete(i)
-
-    def _update_image_listbox(self):
-        self._populate_listbox_from_step_data(self.image_listbox, self.step_data, 'detection_target')
-
-    def _update_fallback_image_listbox(self):
-        fallback_action = self.step_data.get('on_fail', {})
-        self._populate_listbox_from_step_data(self.fallback_image_listbox, fallback_action, 'detection_target')
-
-    def _update_until_image_listbox(self):
-        self._populate_listbox_from_step_data(self.until_image_listbox, self.step_data, 'loop_condition_target')
-
-    def _populate_listbox_from_step_data(self, listbox, data_dict, key):
-        if not data_dict:
-            return
-
-        targets = data_dict.get(key, [])
-        if isinstance(targets, str):
-            # Handle old format where target was a single string path
-            targets = [os.path.basename(targets)]
-
-        listbox.delete(0, tk.END)
-        for target in targets:
-            # The saved value might be a full path, so we take the basename
-            listbox.insert(tk.END, os.path.basename(target))
-
-    def _get_ocr_region_display_text(self):
-        region = self.ocr_region
-        # Defensively check if region is a dictionary and has the required keys
-        if isinstance(region, dict) and all(k in region for k in ['x', 'y', 'width', 'height']):
-            return f"OCR Region: X={region['x']}, Y={region['y']}, W={region['width']}, H={region['height']}"
-        return "Not set. Click 'Set Region' to define the area to read."
-
-    def set_ocr_region(self):
-        self.app.log("Opening region selector for OCR...")
-        selector = RegionSelector(self, self.on_ocr_region_selected)
-        self.wait_window(selector)
-        self.grab_set()
-
-    def on_ocr_region_selected(self, region):
-        self.ocr_region = region
-        self.ocr_region_label_var.set(self._get_ocr_region_display_text())
-        self.app.log("OCR region set.")
-
-
-class ActionPreview(tk.Toplevel):
-    def __init__(self, master, x, y, size=30, duration=1000):
-        super().__init__(master)
-        self.geometry(f"{size}x{size}+{x - size // 2}+{y - size // 2}")
-        self.attributes('-alpha', 0.6)
-        self.attributes('-topmost', True)
-        self.overrideredirect(True) # No title bar or borders
-
-        canvas = tk.Canvas(self, bg='red', highlightthickness=0)
-        canvas.pack(fill="both", expand=True)
-
-        self.after(duration, self.destroy)
-
-class HotkeyChangeDialog(tk.Toplevel):
-    def __init__(self, master):
-        super().__init__(master)
-        self.master = master
-        self.title("Set Hotkey")
-        self.geometry("300x150")
-        self.configure(bg=self.master.bg_color)
-        self.transient(self.master)
-        self.grab_set()
-
-        self.hotkey_str = tk.StringVar(value="Press any key...")
-
-        ttk.Label(self, text="Press any key to set it as the new hotkey.", wraplength=280).pack(pady=10)
-        ttk.Label(self, textvariable=self.hotkey_str, style="Card.TLabel", width=20, font=("Arial", 12), anchor="center").pack(pady=10, ipady=10)
-        ttk.Button(self, text="Cancel", command=self.on_close).pack(pady=10)
-
-        self.listener = keyboard.Listener(on_press=self.on_key_press, suppress=True)
-        self.listener.start()
-
-        self.protocol("WM_DELETE_WINDOW", self.on_close)
-
-    def on_key_press(self, key):
-        try:
-            key_name = None
-            if isinstance(key, keyboard.Key):
-                key_name = key.name
-            elif isinstance(key, keyboard.KeyCode):
-                key_name = key.char
-
-            if key_name:
-                self.hotkey_str.set(key_name)
-                self.master.settings_manager.set_setting('hotkey', key_name)
-                self.master.hotkey_label_var.set(key_name)
-                self.master.log(f"Hotkey changed to: {key_name}")
-                # Restart the main listener if it's running
-                if self.master.hotkey_listener:
-                    self.master.hotkey_listener.stop()
-                    self.master.start_hotkey_listener()
-        except Exception as e:
-            self.master.log(f"Could not set hotkey: {e}")
-        finally:
-            self.after(100, self.on_close)
-
-    def on_close(self):
-        if self.listener.running:
-            self.listener.stop()
-        self.destroy()
-
-if __name__ == "__main__":
-    app = App()
-    app.mainloop()
