+++ pixel_bot/main.py
@@ -1,0 +1,5 @@
+from pixel_bot.gui.main_window import App
+
+if __name__ == "__main__":
+    app = App()
+    app.mainloop()
