+++ pixel_bot/utils/image_analyzer.py
