+++ pixel_bot/utils/settings_manager.py
